leodos_libcfs::cfe::es

Module app

A safe, idiomatic application framework for cFS.

This module provides the App struct as a handle to cFS services and the AppMain trait to define application behavior.

Structs

App

A context handle for the currently running cFS application.

AppId

A type-safe, zero-cost wrapper for a cFE Application ID.

AppInfo

A high-level wrapper around the FFI’s CFE_ES_AppInfo_t.

Traits

AppMain

Defines the behavior of a cFS application.

Functions

default_panic_handler

Provides a default panic handler that logs the panic to the cFE System Log and exits the application. This is highly recommended for all applications.

start

The main entry point and lifecycle manager for a cFS application.