leodos_libcfs::cfe::es::app

Trait AppMain

pub trait AppMain: Sized {
    // Required methods
    fn init() -> Result<Self>;
    fn run_cycle(&mut self) -> Result<()>;
}

Defines the behavior of a cFS application.

Your application’s state struct should implement this trait.

Required Methods

fn init() -> Result<Self>

The initialization routine for the application.

Errors

This function should return an error if initialization fails. The error will be logged to the system log, and the application will exit. This function is called once at startup. It should perform all necessary cFS resource registration (EVS, SB pipes, tables, etc.).

On success, it returns Ok(Self), creating the initial state of your application.

fn run_cycle(&mut self) -> Result<()>

The main processing loop for the application.

This function is called once per cFS scheduler cycle.

Errors

If this function returns an error, the main application loop will terminate, and the application will exit. primary logic of your application, such as reading from a software bus pipe.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.