leodos_libcfs::cfe::es

Module system

Safe wrappers for CFE System-level APIs.

Enums

ResetSubtype

The specific cause of the most recent reset.

ResetType

The type of reset the processor most recently underwent.

SystemState

The overall state of the cFE system, used for startup synchronization.

Functions

background_wakeup

Wakes up the ES background task to process pending jobs.

get_processor_id

Returns the cFE-defined processor ID for the current CPU.

get_reset_type

Returns the type and subtype of the most recent processor reset.

get_spacecraft_id

Returns the cFE-defined spacecraft ID.

process_async_event

Notifies ES that an asynchronous event was detected by the underlying OS/PSP.

reset_cfe

Reset the cFE Core and all cFE Applications. This function does not return.

wait_for_startup_syncDeprecated

Allows an application to wait until all cFE apps have reached the OPERATIONAL state.

wait_for_system_state

Allows an application to wait until the cFE system reaches a minimum state.