leodos_libcfs::cfe::sb::msg

Struct MessageRef

pub struct MessageRef<'a> { /* private fields */ }

A safe, read-only wrapper around a CFE message byte slice.

This provides methods to access header fields without needing unsafe code or raw pointers in your application.

Implementations

impl<'a> MessageRef<'a>

pub fn new(slice: &'a [u8]) -> Self

Creates a new MessageRef from a byte slice.

pub fn as_slice(&self) -> &'a [u8]

Returns the underlying byte slice of the entire message.

pub fn header_version(&self) -> Result<u16>

Gets the header version from the message header.

pub fn apid(&self) -> Result<u16>

Gets the APID from the message header.

pub fn size(&self) -> Result<usize>

Gets the total size of the message from its header.

pub fn msg_id(&self) -> Result<MsgId>

Gets the message ID from the message header.

pub fn fcn_code(&self) -> Result<u16>

Gets the function code from a command message header.

Returns an error if the message does not have a secondary command header.

pub fn time(&self) -> Result<SysTime>

Gets the timestamp from a telemetry message header.

Returns an error if the message does not have a secondary telemetry header.

pub fn sequence_count(&self) -> Result<u16>

Gets the sequence count from the message header.

pub fn validate_checksum(&self) -> Result<bool>

Validates the checksum of a command message.

Returns Ok(true) if the checksum is valid, Ok(false) if it is not. Returns an error if the message does not have a command secondary header.

pub fn get_type(&self) -> Result<MsgType>

Gets the message type (Command or Telemetry) from the header.

pub fn has_secondary_header(&self) -> Result<bool>

Checks if the message header indicates the presence of a secondary header.

pub unsafe fn user_data(&self) -> *mut c_void

Gets a pointer to the user data portion of the message.

C-API Mapping

This is a wrapper for CFE_SB_GetUserData.

Safety

The caller must ensure the returned pointer is cast to the correct payload struct type.

pub fn user_data_length(&self) -> usize

Gets the length of the user data portion of the message.

pub fn segmentation_flag(&self) -> Result<CFE_MSG_SegmentationFlag_t>

Gets the segmentation flag from the message header.

pub fn eds_version(&self) -> Result<u16>

Gets the EDS version from the message header.

pub fn endian(&self) -> Result<CFE_MSG_Endian_t>

Gets the endianness indicator from the message header.

pub fn playback_flag(&self) -> Result<CFE_MSG_PlaybackFlag_t>

Gets the playback flag from the message header.

pub fn subsystem(&self) -> Result<u16>

Gets the subsystem ID from the message header.

pub fn system(&self) -> Result<u16>

Gets the system ID from the message header.