leodos_libcfs::cfe

Module time

CFE Time Services (TIME) interface.

This module provides safe wrappers for the cFE Time Services API, which is the primary source for mission-synchronized time in a cFS system. It handles spacecraft time, Mission Elapsed Time (MET), and conversions between them.

Structs

SysTime

A wrapper around CFE_TIME_SysTime_t representing a specific time.

Functions

external_gps

Provides time from an external source like a GPS receiver.

external_met

Provides the Mission Elapsed Time from an external source.

external_time

Provides time from an external source relative to a known epoch.

external_tone

Provides the 1 Hz signal from an external source.

get_clock_info

Provides information about the spacecraft clock as a bitmask.

get_clock_state

Returns the current state of the spacecraft clock.

get_leap_seconds

Returns the current value of the leap seconds counter.

get_met_seconds

Returns the current seconds count of the mission-elapsed time.

get_met_subseconds

Returns the current sub-seconds count of the mission-elapsed time.

microseconds_to_subseconds

Converts microseconds to a subseconds value.

register_synch_callback

Registers a synchronization callback to be called on time synchronization events.

subseconds_to_microseconds

Converts a subseconds value (1/2^32 seconds) to microseconds.

unregister_synch_callback

Unregisters a previously registered synchronization callback.

Type Aliases

SynchCallback

A type alias for the callback function used for time synchronization events.