leodos_libcfs::error

Enum Error

#[non_exhaustive]
pub enum Error {

    CfeStatusWrongMsgLength,
    CfeStatusUnknownMsgId,
    CfeStatusBadCommandCode,
    CfeStatusExternalResourceFail,
    CfeStatusRequestAlreadyPending,
    CfeStatusValidationFailure,
    CfeStatusRangeError,
    CfeStatusIncorrectState,
    CfeStatusNotImplemented,
    CfeEvsUnknownFilter,
    CfeEvsAppNotRegistered,
    CfeEvsAppIllegalAppId,
    CfeEvsAppFilterOverload,
    CfeEvsResetAreaPointer,
    CfeEvsEvtNotRegistered,
    CfeEvsFileWriteError,
    CfeEvsInvalidParameter,
    CfeEvsAppSquelched,
    CfeEvsNotImplemented,
    CfeEsErrResourceIdNotValid,
    CfeEsErrNameNotFound,
    CfeEsErrAppCreate,
    CfeEsErrChildTaskCreate,
    CfeEsErrSysLogFull,
    CfeEsErrMemBlockSize,
    CfeEsErrLoadLib,
    CfeEsBadArgument,
    CfeEsErrChildTaskRegister,
    CfeEsCdsInsufficientMemory,
    CfeEsCdsInvalidName,
    CfeEsCdsInvalidSize,
    CfeEsCdsInvalid,
    CfeEsCdsAccessError,
    CfeEsFileIoErr,
    CfeEsRstAccessErr,
    CfeEsErrAppRegister,
    CfeEsErrChildTaskDelete,
    CfeEsErrChildTaskDeleteMainTask,
    CfeEsCdsBlockCrcErr,
    CfeEsMutSemDeleteErr,
    CfeEsBinSemDeleteErr,
    CfeEsCountSemDeleteErr,
    CfeEsQueueDeleteErr,
    CfeEsFileCloseErr,
    CfeEsCdsWrongTypeErr,
    CfeEsCdsOwnerActiveErr,
    CfeEsAppCleanupErr,
    CfeEsTimerDeleteErr,
    CfeEsBufferNotInPool,
    CfeEsTaskDeleteErr,
    CfeEsOperationTimedOut,
    CfeEsNoResourceIdsAvailable,
    CfeEsPoolBlockInvalid,
    CfeEsErrDuplicateName,
    CfeEsNotImplemented,
    CfeFsBadArgument,
    CfeFsInvalidPath,
    CfeFsFnameTooLong,
    CfeFsNotImplemented,
    CfeSbTimeOut,
    CfeSbNoMessage,
    CfeSbBadArgument,
    CfeSbMaxPipesMet,
    CfeSbPipeCrErr,
    CfeSbPipeRdErr,
    CfeSbMsgTooBig,
    CfeSbBufAlocErr,
    CfeSbMaxMsgsMet,
    CfeSbMaxDestsMet,
    SbInternalErr,
    SbWrongMsgType,
    SbBufferInvalid,
    SbNotImplemented,
    CfeTblErrInvalidHandle,
    CfeTblErrInvalidName,
    CfeTblErrInvalidSize,
    CfeTblErrNeverLoaded,
    CfeTblErrRegistryFull,
    CfeTblErrNoAccess,
    CfeTblErrUnregistered,
    CfeTblErrHandlesFull,
    CfeTblErrDuplicateDiffSize,
    CfeTblErrDuplicateNotOwned,
    CfeTblErrNoBufferAvail,
    CfeTblErrDumpOnly,
    CfeTblErrIllegalSrcType,
    CfeTblErrLoadInProgress,
    CfeTblErrFileTooLarge,
    CfeTblErrBadContentId,
    CfeTblErrBadSubtypeId,
    CfeTblErrFileSizeInconsistent,
    CfeTblErrNoStdHeader,
    CfeTblErrNoTblHeader,
    CfeTblErrFilenameTooLong,
    CfeTblErrFileForWrongTable,
    CfeTblErrLoadIncomplete,
    CfeTblErrPartialLoad,
    CfeTblErrInvalidOptions,
    CfeTblErrBadSpacecraftId,
    CfeTblErrBadProcessorId,
    CfeTblMessageError,
    CfeTblErrShortFile,
    CfeTblErrAccess,
    CfeTblBadArgument,
    CfeTblNotImplemented,
    CfeTimeNotImplemented,
    CfeTimeInternalOnly,
    CfeTimeOutOfRange,
    CfeTimeTooManySynchCallbacks,
    CfeTimeCallbackNotRegistered,
    CfeTimeBadArgument,
    OsError,
    OsInvalidPointer,
    OsErrorAddressMisaligned,
    OsErrorTimeout,
    OsInvalidIntNum,
    OsSemFailure,
    OsSemTimeout,
    OsQueueEmpty,
    OsQueueFull,
    OsQueueTimeout,
    OsQueueInvalidSize,
    OsQueueIdError,
    OsErrNameTooLong,
    OsErrNoFreeIds,
    OsErrNameTaken,
    OsErrInvalidId,
    OsErrNameNotFound,
    OsErrSemNotFull,
    OsErrInvalidPriority,
    OsInvalidSemValue,
    OsErrFile,
    OsErrNotImplemented,
    OsTimerErrInvalidArgs,
    OsTimerErrTimerId,
    OsTimerErrUnavailable,
    OsTimerErrInternal,
    OsErrObjectInUse,
    OsErrBadAddress,
    OsErrIncorrectObjState,
    OsErrIncorrectObjType,
    OsErrStreamDisconnected,
    OsErrOperationNotSupported,
    OsErrInvalidSize,
    OsErrOutputTooLarge,
    OsErrInvalidArgument,
    OsFsErrPathTooLong,
    OsFsErrNameTooLong,
    OsFsErrDriveNotCreated,
    OsFsErrDeviceNotFree,
    OsFsErrPathInvalid,
    InvalidString,
    Unhandled(i32),

}

Represents all possible errors and status codes from the CFE and OSAL APIs.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

CfeStatusWrongMsgLength

Wrong Message Length.

CfeStatusUnknownMsgId

Unknown Message ID.

CfeStatusBadCommandCode

Bad Command Code.

CfeStatusExternalResourceFail

External resource failure.

CfeStatusRequestAlreadyPending

Request already pending.

CfeStatusValidationFailure

Request or input value failed basic structural validation.

CfeStatusRangeError

Request or input value is out of range.

CfeStatusIncorrectState

Cannot process request at this time.

CfeStatusNotImplemented

Not Implemented.

CfeEvsUnknownFilter

Unknown Filter scheme.

CfeEvsAppNotRegistered

Application not registered with EVS.

CfeEvsAppIllegalAppId

Illegal Application ID.

CfeEvsAppFilterOverload

Application event filter overload.

CfeEvsResetAreaPointer

Could not get pointer to the ES Reset area.

CfeEvsEvtNotRegistered

EventID argument was not found in any registered event filter.

CfeEvsFileWriteError

A file write error occurred while processing an EVS command.

CfeEvsInvalidParameter

Invalid parameter supplied to EVS command.

CfeEvsAppSquelched

Event squelched due to being sent at too high a rate.

CfeEvsNotImplemented

EVS feature is not implemented.

CfeEsErrResourceIdNotValid

Resource ID is not valid.

CfeEsErrNameNotFound

Resource Name not found.

CfeEsErrAppCreate

Error loading or creating an App.

CfeEsErrChildTaskCreate

Error creating a child task.

CfeEsErrSysLogFull

The cFE system Log is full.

CfeEsErrMemBlockSize

The memory block size requested is invalid.

CfeEsErrLoadLib

Could not load the shared library.

CfeEsBadArgument

Bad parameter passed into an ES API.

CfeEsErrChildTaskRegister

Errors occurred when trying to register a child task.

CfeEsCdsInsufficientMemory

CDS is larger than the remaining CDS memory.

CfeEsCdsInvalidName

CDS name is too long or empty.

CfeEsCdsInvalidSize

CDS size is beyond the applicable limits.

CfeEsCdsInvalid

The CDS contents are invalid.

CfeEsCdsAccessError

The CDS was inaccessible.

CfeEsFileIoErr

A file operation failed.

CfeEsRstAccessErr

BSP failed to return the reset area address.

CfeEsErrAppRegister

A task cannot be registered in ES global tables.

CfeEsErrChildTaskDelete

Error deleting a child task.

CfeEsErrChildTaskDeleteMainTask

Attempted to delete a cFE App Main Task with DeleteChildTask API.

CfeEsCdsBlockCrcErr

CDS Data Block CRC does not match stored CRC.

CfeEsMutSemDeleteErr

Error deleting a Mutex Semaphore during task cleanup.

CfeEsBinSemDeleteErr

Error deleting a Binary Semaphore during task cleanup.

CfeEsCountSemDeleteErr

Error deleting a Counting Semaphore during task cleanup.

CfeEsQueueDeleteErr

Error deleting a Queue during task cleanup.

CfeEsFileCloseErr

Error closing a file during task cleanup.

CfeEsCdsWrongTypeErr

CDS is not the correct type for the operation.

CfeEsCdsOwnerActiveErr

Attempted to delete a CDS while its owner application is still active.

CfeEsAppCleanupErr

An error occurred during application cleanup.

CfeEsTimerDeleteErr

Error deleting a Timer during task cleanup.

CfeEsBufferNotInPool

The specified address is not in the memory pool.

CfeEsTaskDeleteErr

Error deleting a task during cleanup.

CfeEsOperationTimedOut

The timeout for a given operation was exceeded.

CfeEsNoResourceIdsAvailable

Maximum number of resource identifiers has been reached.

CfeEsPoolBlockInvalid

Attempted to “put” a block back into a pool which does not belong to it.

CfeEsErrDuplicateName

Resource creation failed due to the name already existing.

CfeEsNotImplemented

ES feature is not implemented.

CfeFsBadArgument

A parameter given to a File Services API did not pass validation.

CfeFsInvalidPath

FS was unable to extract a filename from a path string.

CfeFsFnameTooLong

FS filename string is too long.

CfeFsNotImplemented

FS feature is not implemented.

CfeSbTimeOut

A pipe receive operation timed out.

CfeSbNoMessage

A pipe was polled but contained no message.

CfeSbBadArgument

A parameter given to a Software Bus API did not pass validation.

CfeSbMaxPipesMet

The maximum number of pipes are already in use.

CfeSbPipeCrErr

The underlying OS queue for a pipe could not be created.

CfeSbPipeRdErr

An error occurred at the OS queue read level.

CfeSbMsgTooBig

The message size exceeds the maximum allowed SB message size.

CfeSbBufAlocErr

The SB message buffer pool has been depleted.

CfeSbMaxMsgsMet

The SB routing table cannot accommodate another unique message ID.

CfeSbMaxDestsMet

The SB routing table cannot accommodate another destination for a message ID.

SbInternalErr

An internal SB index is out of range.

SbWrongMsgType

A message header operation was requested on a message of the wrong type.

SbBufferInvalid

A request to release or send a zero-copy buffer is invalid.

SbNotImplemented

SB feature is not implemented.

CfeTblErrInvalidHandle

The provided table handle is not valid.

CfeTblErrInvalidName

The provided table name is not valid.

CfeTblErrInvalidSize

The provided table size is not valid.

CfeTblErrNeverLoaded

The table has not yet been loaded with data.

CfeTblErrRegistryFull

The table registry is full.

CfeTblErrNoAccess

The application does not have access to the table.

CfeTblErrUnregistered

The application is trying to access an unregistered table.

CfeTblErrHandlesFull

The table handle array is full.

CfeTblErrDuplicateDiffSize

An app tried to register a table with the same name but a different size.

CfeTblErrDuplicateNotOwned

An app tried to register a table owned by a different application.

CfeTblErrNoBufferAvail

No working buffer was available for a table load.

CfeTblErrDumpOnly

A load was attempted on a “Dump Only” table.

CfeTblErrIllegalSrcType

The source type for a table load was illegal.

CfeTblErrLoadInProgress

A table load was attempted while another load was in progress.

CfeTblErrFileTooLarge

The table file is larger than the table’s buffer.

CfeTblErrBadContentId

The table file’s content ID was not that of a table image.

CfeTblErrBadSubtypeId

The table file’s Subtype ID was not a table image file.

CfeTblErrFileSizeInconsistent

The table file’s size is inconsistent with its header.

CfeTblErrNoStdHeader

The table file’s standard cFE File Header was invalid.

CfeTblErrNoTblHeader

The table file’s cFE Table File Header was invalid.

CfeTblErrFilenameTooLong

The filename for a table load was too long.

CfeTblErrFileForWrongTable

The table file header indicates it is for a different table.

CfeTblErrLoadIncomplete

The table file load was larger than what was read from the file.

CfeTblErrPartialLoad

A partial load was attempted on an uninitialized table.

CfeTblErrInvalidOptions

An illegal combination of table options was used.

CfeTblErrBadSpacecraftId

The table file failed validation for Spacecraft ID.

CfeTblErrBadProcessorId

The table file failed validation for Processor ID.

CfeTblMessageError

The TBL command was not processed successfully.

CfeTblErrShortFile

The TBL file is shorter than indicated in the file header.

CfeTblErrAccess

The TBL file could not be opened by the OS.

CfeTblBadArgument

A parameter given to a Table API did not pass validation.

CfeTblNotImplemented

TBL feature is not implemented.

CfeTimeNotImplemented

TIME feature is not implemented.

CfeTimeInternalOnly

TIME Services is commanded to not accept external time data.

CfeTimeOutOfRange

New time data from an external source is invalid.

CfeTimeTooManySynchCallbacks

An attempt to register too many Time Services Synchronization callbacks was made.

CfeTimeCallbackNotRegistered

The specified callback function was not in the Synchronization Callback Registry.

CfeTimeBadArgument

A parameter given to a TIME Services API did not pass validation.

OsError

Failed execution.

OsInvalidPointer

Invalid pointer.

OsErrorAddressMisaligned

Address misalignment.

OsErrorTimeout

Timeout.

OsInvalidIntNum

Invalid Interrupt number.

OsSemFailure

Semaphore failure.

OsSemTimeout

Semaphore timeout.

OsQueueEmpty

Queue empty.

OsQueueFull

Queue full.

OsQueueTimeout

Queue timeout.

OsQueueInvalidSize

Queue invalid size.

OsQueueIdError

Queue ID error.

OsErrNameTooLong

Name length too long.

OsErrNoFreeIds

No free IDs.

OsErrNameTaken

Name taken.

OsErrInvalidId

Invalid ID.

OsErrNameNotFound

Name not found.

OsErrSemNotFull

Semaphore not full.

OsErrInvalidPriority

Invalid priority.

OsInvalidSemValue

Invalid semaphore value.

OsErrFile

Generic file error.

OsErrNotImplemented

Not implemented.

OsTimerErrInvalidArgs

Timer invalid arguments.

OsTimerErrTimerId

Timer ID error.

OsTimerErrUnavailable

Timer unavailable.

OsTimerErrInternal

Timer internal error.

OsErrObjectInUse

Object in use.

OsErrBadAddress

Bad address.

OsErrIncorrectObjState

Incorrect object state.

OsErrIncorrectObjType

Incorrect object type.

OsErrStreamDisconnected

Stream disconnected.

OsErrOperationNotSupported

Requested operation not supported on supplied object(s).

OsErrInvalidSize

Invalid Size.

OsErrOutputTooLarge

Size of output exceeds limit.

OsErrInvalidArgument

Invalid argument value.

OsFsErrPathTooLong

FS path too long.

OsFsErrNameTooLong

FS name too long.

OsFsErrDriveNotCreated

FS drive not created.

OsFsErrDeviceNotFree

FS device not free.

OsFsErrPathInvalid

FS path invalid.

InvalidString

A string passed to the API contained an interior null character.

Unhandled(i32)

An unhandled or unknown CFE/OSAL/PSP error code. This may indicate a new or platform-specific error code not yet included in this enum.