leodos_libcfs::ffi

Function CFE_ES_CalculateCRC

pub unsafe extern "C" fn CFE_ES_CalculateCRC(
    DataPtr: *const c_void,
    DataLength: usize,
    InputCRC: uint32,
    TypeCRC: CFE_ES_CrcType_Enum_t,
) -> uint32

/ /** \brief Calculate a CRC on a block of memory

\par Description This routine calculates a cyclic redundancy check (CRC) on a block of memory. The CRC algorithm used is determined by the last parameter.

\par Assumptions, External Events, and Notes: None

\param[in] DataPtr Pointer to the base of the memory block.

\param[in] DataLength The number of bytes in the memory block.

\param[in] InputCRC A starting value for use in the CRC calculation. This parameter allows the user to calculate the CRC of non-contiguous blocks as a single value. Nominally, the user should set this value to zero.

\param[in] TypeCRC One of the following CRC algorithm selections defined in CFE_ES_CrcType_Enum_t

\return The result of the CRC calculation on the specified memory block. If the TypeCRC is unimplemented will return 0. If DataPtr is null or DataLength is 0, will return InputCRC