leodos_libcfs::ffi

Function CFE_ES_CreateChildTask

pub unsafe extern "C" fn CFE_ES_CreateChildTask(
    TaskIdPtr: *mut CFE_ES_TaskId_t,
    TaskName: *const c_char,
    FunctionPtr: CFE_ES_ChildTaskMainFuncPtr_t,
    StackPtr: CFE_ES_StackPointer_t,
    StackSize: usize,
    Priority: CFE_ES_TaskPriority_Atom_t,
    Flags: uint32,
) -> CFE_Status_t

/ /** \brief Creates a new task under an existing Application

\par Description This routine creates a new task (a separate execution thread) owned by the calling Application.

\par Assumptions, External Events, and Notes: None

\param[out] TaskIdPtr A pointer to a variable that will be filled in with the new task’s ID @nonnull. TaskIdPtr is the Task ID of the newly created child task.

\param[in] TaskName A pointer to a string containing the desired name of the new task @nonnull. This can be up to #OS_MAX_API_NAME characters, including the trailing null.

\param[in] FunctionPtr A pointer to the function that will be spawned as a new task @nonnull.

\param[in] StackPtr A pointer to the location where the child task’s stack pointer should start. NOTE: Not all underlying operating systems support this parameter. The CFE_ES_TASK_STACK_ALLOCATE constant may be passed to indicate that the stack should be dynamically allocated.

\param[in] StackSize The number of bytes to allocate for the new task’s stack @nonzero.

\param[in] Priority The priority for the new task. Lower numbers are higher priority, with 0 being the highest priority.

\param[in] Flags Reserved for future expansion.

\return Execution status, see \ref CFEReturnCodes \retval #CFE_SUCCESS \copybrief CFE_SUCCESS \retval #CFE_ES_ERR_CHILD_TASK_CREATE \copybrief CFE_ES_ERR_CHILD_TASK_CREATE \retval #CFE_ES_BAD_ARGUMENT \copybrief CFE_ES_BAD_ARGUMENT \retval #CFE_ES_ERR_RESOURCEID_NOT_VALID \copybrief CFE_ES_ERR_RESOURCEID_NOT_VALID

\sa #CFE_ES_DeleteChildTask, #CFE_ES_ExitChildTask