leodos_libcfs::ffi

Function CFE_ES_GetPoolBuf

pub unsafe extern "C" fn CFE_ES_GetPoolBuf(
    BufPtr: *mut CFE_ES_MemPoolBuf_t,
    Handle: CFE_ES_MemHandle_t,
    Size: usize,
) -> int32

/ /** \brief Gets a buffer from the memory pool created by #CFE_ES_PoolCreate or #CFE_ES_PoolCreateNoSem

\par Description This routine obtains a block of memory from the memory pool supplied by the calling application.

\par Assumptions, External Events, and Notes: -# The size allocated from the memory pool is, at a minimum, 12 bytes more than requested.

\param[out] BufPtr A pointer to the Application’s pointer @nonnull in which will be stored the address of the allocated memory buffer. *BufPtr is the address of the requested buffer.

\param[in] Handle The handle to the memory pool as returned by #CFE_ES_PoolCreate or #CFE_ES_PoolCreateNoSem.

\param[in] Size The size of the buffer requested. NOTE: The size allocated may be larger.

\return Bytes Allocated, or error code \ref CFEReturnCodes \retval #CFE_ES_ERR_RESOURCEID_NOT_VALID \copybrief CFE_ES_ERR_RESOURCEID_NOT_VALID \retval #CFE_ES_ERR_MEM_BLOCK_SIZE \copybrief CFE_ES_ERR_MEM_BLOCK_SIZE \retval #CFE_ES_BAD_ARGUMENT \copybrief CFE_ES_BAD_ARGUMENT

\sa #CFE_ES_PoolCreate, #CFE_ES_PoolCreateNoSem, #CFE_ES_PoolCreateEx, #CFE_ES_PutPoolBuf, #CFE_ES_GetMemPoolStats, #CFE_ES_GetPoolBufInfo