leodos_libcfs::ffi

Function CFE_ES_GetResetType

pub unsafe extern "C" fn CFE_ES_GetResetType(
    ResetSubtypePtr: *mut uint32,
) -> int32

/ /** \brief Return the most recent Reset Type

\par Description Provides the caller with codes that identifies the type of Reset the processor most recently underwent. The caller can also obtain information on what caused the reset by supplying a pointer to a variable that will be filled with the Reset Sub-Type.

\par Assumptions, External Events, and Notes: None

\param[in, out] ResetSubtypePtr Pointer to \c uint32 type variable in which the Reset Sub-Type will be stored. The caller can set this pointer to NULL if the Sub-Type is of no interest. \n ResetSubtypePtr If the provided pointer was not \c NULL, the Reset Sub-Type is stored at the given address. For a list of possible Sub-Type values, see \link #CFE_PSP_RST_SUBTYPE_POWER_CYCLE “Reset Sub-Types” \endlink.

\return Processor reset type \retval #CFE_PSP_RST_TYPE_POWERON \copybrief CFE_PSP_RST_TYPE_POWERON \retval #CFE_PSP_RST_TYPE_PROCESSOR \copybrief CFE_PSP_RST_TYPE_PROCESSOR

\sa #CFE_ES_GetAppID, #CFE_ES_GetAppIDByName, #CFE_ES_GetAppName, #CFE_ES_GetTaskInfo