leodos_libcfs::ffi

Function CFE_ES_GetTaskInfo

pub unsafe extern "C" fn CFE_ES_GetTaskInfo(
    TaskInfo: *mut CFE_ES_TaskInfo_t,
    TaskId: CFE_ES_TaskId_t,
) -> CFE_Status_t

/ /** \brief Get Task Information given a specified Task ID

\par Description This routine retrieves the information about a Task associated with a specified Task ID. The information includes Task Name, and Parent/Creator Application ID.

\par Assumptions, External Events, and Notes: None

\param[out] TaskInfo Pointer to a \c CFE_ES_TaskInfo_t structure @nonnull that holds the specific task information. *TaskInfo is the filled out \c CFE_ES_TaskInfo_t structure containing the Task Name, Parent App Name, Parent App ID among other fields.

\param[in] TaskId Application ID of Application whose name is being requested.

\return Execution status, see \ref CFEReturnCodes \retval #CFE_SUCCESS \copybrief CFE_SUCCESS \retval #CFE_ES_ERR_RESOURCEID_NOT_VALID \copybrief CFE_ES_ERR_RESOURCEID_NOT_VALID \retval #CFE_ES_BAD_ARGUMENT \copybrief CFE_ES_BAD_ARGUMENT

\sa #CFE_ES_GetTaskID, #CFE_ES_GetTaskIDByName, #CFE_ES_GetTaskName