leodos_libcfs::ffi

Function CFE_ES_Main

pub unsafe extern "C" fn CFE_ES_Main(
    StartType: uint32,
    StartSubtype: uint32,
    ModeId: uint32,
    StartFilePath: *const c_char,
)

/ /** \brief cFE Main Entry Point used by Board Support Package to start cFE

\par Description cFE main entry point. This is the entry point into the cFE software. It is called only by the Board Support Package software.

\par Assumptions, External Events, and Notes: None

\param[in] StartType Identifies whether this was a #CFE_PSP_RST_TYPE_POWERON or #CFE_PSP_RST_TYPE_PROCESSOR.

\param[in] StartSubtype Specifies, in more detail, what caused the \c StartType identified above. See #CFE_PSP_RST_SUBTYPE_POWER_CYCLE for possible examples.

\param[in] ModeId Identifies the source of the Boot as determined by the BSP.

\param[in] StartFilePath Identifies the startup file to use to initialize the cFE apps.

\sa #CFE_ES_ResetCFE