leodos_libcfs::ffi

Function CFE_ES_PerfLogAdd

pub unsafe extern "C" fn CFE_ES_PerfLogAdd(
    Marker: uint32,
    EntryExit: uint32,
)

/ /** \brief Adds a new entry to the data buffer

Function called by #CFE_ES_PerfLogEntry and #CFE_ES_PerfLogExit macros

\par Description This function logs the entry and exit marker for the specified \c id. This function is used by the Software Performance Analysis tool.

\par Assumptions, External Events, and Notes:

\c Marker limited to the range of 0 to #CFE_MISSION_ES_PERF_MAX_IDS - 1. Any performance ids outside of this range will be ignored and will be flagged as an error.

This function implements a circular buffer using an array. DataStart points to first stored entry DataEnd points to next available entry if DataStart == DataEnd then the buffer is either empty or full depending on the value of the DataCount

Time is stored as 2 32 bit integers, (TimerLower32, TimerUpper32): TimerLower32 is the current value of the hardware timer register. TimerUpper32 is the number of times the timer has rolled over.

\param[in] Marker Identifier of the specific event or marker. \param[in] EntryExit Used to specify Entry(0) or Exit(1)

\sa #CFE_ES_PerfLogEntry, #CFE_ES_PerfLogExit