leodos_libcfs::ffi

Function CFE_ES_RegisterCDS

pub unsafe extern "C" fn CFE_ES_RegisterCDS(
    CDSHandlePtr: *mut CFE_ES_CDSHandle_t,
    BlockSize: usize,
    Name: *const c_char,
) -> CFE_Status_t

/ /** \brief Reserve space (or re-obtain previously reserved space) in the Critical Data Store (CDS)

\par Description This routine allocates a block of memory in the Critical Data Store and associates it with the calling Application. The memory can survive an Application restart as well as a Processor Reset.

\par Assumptions, External Events, and Notes: This function does not clear or otherwise initialize/modify the data within the CDS block. If this function returns #CFE_ES_CDS_ALREADY_EXISTS the block may already have valid data in it.

    If a new CDS block is reserved (either because the name did not exist, or existed as a different
    size) it is the responsibility of the calling application to fill the CDS block with valid data.
    This is indicated by a #CFE_SUCCESS return code, and in this case the calling application should
    ensure that it also calls CFE_ES_CopyToCDS() to fill the block with valid data.

\param[out] CDSHandlePtr Pointer Application’s variable that will contain the CDS Memory Block Handle @nonnull. HandlePtr is the handle of the CDS block that can be used in #CFE_ES_CopyToCDS and #CFE_ES_RestoreFromCDS.

\param[in] BlockSize The number of bytes needed in the CDS @nonzero.

\param[in] Name A pointer to a character string @nonnull containing an application unique name of #CFE_MISSION_ES_CDS_MAX_NAME_LENGTH characters or less.

\return Execution status, see \ref CFEReturnCodes \retval #CFE_SUCCESS The memory block was successfully created in the CDS. \retval #CFE_ES_NOT_IMPLEMENTED The processor does not support a Critical Data Store. \retval #CFE_ES_CDS_ALREADY_EXISTS \copybrief CFE_ES_CDS_ALREADY_EXISTS \retval #CFE_ES_CDS_INVALID_SIZE \copybrief CFE_ES_CDS_INVALID_SIZE \retval #CFE_ES_CDS_INVALID_NAME \copybrief CFE_ES_CDS_INVALID_NAME \retval #CFE_ES_BAD_ARGUMENT \copybrief CFE_ES_BAD_ARGUMENT \retval #CFE_ES_CDS_INVALID \covtest \copybrief CFE_ES_CDS_INVALID

\sa #CFE_ES_CopyToCDS, #CFE_ES_RestoreFromCDS