leodos_libcfs::ffi

Function CFE_ES_WaitForStartupSync

pub unsafe extern "C" fn CFE_ES_WaitForStartupSync(
    TimeOutMilliseconds: uint32,
)

/ /** \brief Allow an Application to Wait for the “OPERATIONAL” global system state

\par Description This is the API that allows an app to wait for the rest of the apps to complete their entire initialization before continuing. It is most useful for applications such as Health and Safety or the Scheduler that need to wait until applications exist and are running before sending out packets to them.

    This is a specialized wrapper for CFE_ES_WaitForSystemState for compatibility
    with applications using this API.

\par Assumptions, External Events, and Notes: This API should only be called as the last item of an Apps initialization. In addition, this API should only be called by an App that is started from the ES Startup file. It should not be used by an App that is started after the system is running. ( Although it will cause no harm )

\param[in] TimeOutMilliseconds The timeout value in Milliseconds. This parameter must be at least 1000. Lower values will be rounded up. There is not an option to wait indefinitely to avoid hanging a critical application because a non-critical app did not start.

\sa #CFE_ES_RunLoop