leodos_libcfs::ffi

Function CFE_EVS_Register

pub unsafe extern "C" fn CFE_EVS_Register(
    Filters: *const c_void,
    NumEventFilters: uint16,
    FilterScheme: uint16,
) -> CFE_Status_t

\brief Register an application for receiving event services

\par Description This routine registers an application with event services and allocates/initializes the internal data structures used to support this application’s events. An application may not send events unless it has called this routine. The routine also accepts a filter array structure for applications requiring event filtering. In the current implementation of the EVS, only the binary filtering scheme is supported. See section TBD of the cFE Application Programmer’s Guide for a description of the behavior of binary filters. Applications may call #CFE_EVS_Register more than once, but each call will wipe out all filters registered by previous calls (filter registration is NOT cumulative).

\par Assumptions, External Events, and Notes: Note: Event filters can be added, deleted or modified by ground commands. All filtering schemes include a default setting that results in no filtering (such as #CFE_EVS_NO_FILTER for binary filters).

     <b> Filter Scheme: </b> Binary <BR>
     <b> Code: </b> CFE_EVS_EventFilter_BINARY <BR>
     <b> Filter Structure: </b>
     \code
                typedef struct CFE_EVS_BinFilter {
                      uint16  EventID,
                      uint16  Mask ;
                } CFE_EVS_BinFilter_t;
     \endcode

\param[in] Filters Pointer to an array of event message filters, or NULL if no filtering is desired. The structure of an event message filter depends on the FilterScheme selected. (see Filter Schemes mentioned above)

\param[in] NumEventFilters The number of event message filters included in this call. This must be less than or equal to the maximum number of events allowed per application (#CFE_PLATFORM_EVS_MAX_EVENT_FILTERS).

\param[in] FilterScheme The event filtering scheme that this application will use. For the first implementation of the event services, only filter type #CFE_EVS_EventFilter_BINARY will be supported.

\return Execution status below or from #CFE_ES_GetAppID, see \ref CFEReturnCodes \retval #CFE_SUCCESS \copybrief CFE_SUCCESS \retval #CFE_EVS_APP_FILTER_OVERLOAD \copybrief CFE_EVS_APP_FILTER_OVERLOAD \retval #CFE_EVS_UNKNOWN_FILTER \copybrief CFE_EVS_UNKNOWN_FILTER \retval #CFE_EVS_APP_ILLEGAL_APP_ID \copybrief CFE_EVS_APP_ILLEGAL_APP_ID \retval #CFE_ES_BAD_ARGUMENT \copybrief CFE_ES_BAD_ARGUMENT