leodos_libcfs::ffi

Function CFE_EVS_SendEventWithAppID

pub unsafe extern "C" fn CFE_EVS_SendEventWithAppID(
    EventID: uint16,
    EventType: CFE_EVS_EventType_Enum_t,
    AppID: CFE_ES_AppId_t,
    Spec: *const c_char,
    ...
) -> CFE_Status_t

\brief Generate a software event given the specified Application ID.

\par Description This routine generates a software event message. If the EventID is not filtered, the event will be sent as a software bus message, optionally logged in the local event log, and optionally sent as an ASCII text string out the enabled output port(s). Note that this function should really only be used from within an API in order to preserve the context of an Application’s event. In general, #CFE_EVS_SendEvent should be used.

\par Assumptions, External Events, and Notes: The Application ID must correspond to a registered application or core service. For messages outside the context of a registered application (for example early in app initialization or if registration fails) #CFE_ES_WriteToSysLog can be used for reporting.

\param[in] EventID A numeric literal used to uniquely identify an application event. The \c EventID is defined and supplied by the application sending the event.

\param[in] EventType A numeric literal used to classify an event, one of: \arg #CFE_EVS_EventType_DEBUG \arg #CFE_EVS_EventType_INFORMATION \arg #CFE_EVS_EventType_ERROR \arg #CFE_EVS_EventType_CRITICAL

\param[in] AppID The Application ID from which the event message should appear.

\param[in] Spec A pointer to a null terminated text string @nonnull describing the output format for the event. This is the same type of format string used for the ANSI \c printf function. Nominally the post-conversion string is limited to 80 characters, but this limit is configurable through the parameter #CFE_MISSION_EVS_MAX_MESSAGE_LENGTH. Characters beyond the limit will be truncated. Do not use floating point conversions (%f, %e, %E, %g, and %G) in the format string unless your application will be running in a system that supports floating point arithmetic. Do not use non-printable characters (\t, \n, etc.) in the format string; they will mess up the formatting when the events are displayed on the ground system.

\return Execution status, see \ref CFEReturnCodes \retval #CFE_SUCCESS \copybrief CFE_SUCCESS \retval #CFE_EVS_APP_NOT_REGISTERED \copybrief CFE_EVS_APP_NOT_REGISTERED \retval #CFE_EVS_APP_ILLEGAL_APP_ID \copybrief CFE_EVS_APP_ILLEGAL_APP_ID \retval #CFE_EVS_INVALID_PARAMETER \copybrief CFE_EVS_INVALID_PARAMETER

\sa #CFE_EVS_SendEvent, #CFE_EVS_SendTimedEvent