leodos_libcfs::ffi

Function CFE_FS_BackgroundFileDumpRequest

pub unsafe extern "C" fn CFE_FS_BackgroundFileDumpRequest(
    Meta: *mut CFE_FS_FileWriteMetaData_t,
) -> int32

/ /** \brief Register a background file dump request

\par Description Puts the previously-initialized metadata into the pending request queue

\par Assumptions, External Events, and Notes: Metadata structure should be stored in a persistent memory area (not on stack) as it must remain accessible by the file writer task throughout the asynchronous job operation.

\param[inout] Meta The background file write persistent state object @nonnull

\return Execution status, see \ref CFEReturnCodes \retval #CFE_FS_BAD_ARGUMENT \copybrief CFE_FS_BAD_ARGUMENT \retval #CFE_FS_INVALID_PATH \copybrief CFE_FS_INVALID_PATH \retval #CFE_STATUS_REQUEST_ALREADY_PENDING \copybrief CFE_STATUS_REQUEST_ALREADY_PENDING \retval #CFE_SUCCESS \copybrief CFE_SUCCESS