leodos_libcfs::ffi

Function CFE_FS_ParseInputFileName

pub unsafe extern "C" fn CFE_FS_ParseInputFileName(
    OutputBuffer: *mut c_char,
    InputName: *const c_char,
    OutputBufSize: usize,
    FileCategory: CFE_FS_FileCategory_t,
) -> int32

/ /** \brief Parse a filename string from the user into a local buffer

\par Description Simplified API for CFE_FS_ParseInputFileNameEx() where input is always known to be a non-empty, null terminated string and the fixed-length input buffer not needed. For instance this may be used where the input is a fixed string from cfe_platform_cfg.h or similar.

\par Assumptions, External Events, and Notes: The parameters are organized such that this is basically like strncpy() with an extra argument, and existing file name accesses which use a direct copy can easily change to use this instead.

\sa CFE_FS_ParseInputFileNameEx()

\param[out] OutputBuffer Buffer to store result @nonnull. \param[in] InputName A null terminated input string @nonnull. \param[in] OutputBufSize Maximum Size of output buffer @nonzero. \param[in] FileCategory The generalized category of file (implies default path/extension)

\return Execution status, see \ref CFEReturnCodes