leodos_libcfs::ffi

Function CFE_FS_ParseInputFileNameEx

pub unsafe extern "C" fn CFE_FS_ParseInputFileNameEx(
    OutputBuffer: *mut c_char,
    InputBuffer: *const c_char,
    OutputBufSize: usize,
    InputBufSize: usize,
    DefaultInput: *const c_char,
    DefaultPath: *const c_char,
    DefaultExtension: *const c_char,
) -> int32

/ /** \brief Parse a filename input from an input buffer into a local buffer

\par Description This provides a more user friendly way to specify file names, using default values for the path and extension, which can vary from system to system.

    If InputBuffer is null or its length is zero, then DefaultInput
    is used as if it was the content of the input buffer.

    If either the pathname or extension is missing from the input,
    it will be added from defaults, with the complete fully-qualified
    filename stored in the output buffer.

\par Assumptions, External Events, and Notes: -# The paths and filenames used here are the standard unix style filenames separated by “/” (path) and “.” (extension) characters. -# Input Buffer has a fixed max length. Parsing will not exceed InputBufSize, and does not need to be null terminated. However parsing will stop at the first null char, when the input is shorter than the maximum.

\param[out] OutputBuffer Buffer to store result @nonnull. \param[in] InputBuffer A input buffer that may contain a file name (e.g. from command) @nonnull. \param[in] OutputBufSize Maximum Size of output buffer @nonzero. \param[in] InputBufSize Maximum Size of input buffer. \param[in] DefaultInput Default value to use for input if InputBffer is empty \param[in] DefaultPath Default value to use for pathname if omitted from input \param[in] DefaultExtension Default value to use for extension if omitted from input

\return Execution status, see \ref CFEReturnCodes \retval #CFE_FS_BAD_ARGUMENT \copybrief CFE_FS_BAD_ARGUMENT \retval #CFE_FS_FNAME_TOO_LONG \copybrief CFE_FS_FNAME_TOO_LONG \retval #CFE_FS_INVALID_PATH \copybrief CFE_FS_INVALID_PATH \retval #CFE_SUCCESS \copybrief CFE_SUCCESS