leodos_libcfs::ffi

Function CFE_FS_SetTimestamp

pub unsafe extern "C" fn CFE_FS_SetTimestamp(
    FileDes: osal_id_t,
    NewTimestamp: CFE_TIME_SysTime_t,
) -> CFE_Status_t

/ /** \brief Modifies the Time Stamp field in the Standard cFE File Header for the specified file

\par Description This API will modify the \link #CFE_FS_Header_t::TimeSeconds timestamp \endlink found in the Standard cFE File Header of the specified file. The timestamp will be replaced with the time specified by the caller.

\par Assumptions, External Events, and Notes: -# The File has already been successfully opened using #OS_OpenCreate and the caller has a legitimate File Descriptor. -# The \c NewTimestamp field has been filled appropriately by the Application. -# File offset behavior: Agnostic on entry since it will move the offset, on success the offset will be at the end of the time stamp, undefined offset behavior for error cases.

\param[in] FileDes File Descriptor obtained from a previous call to #OS_OpenCreate that is associated with the file whose header is to be read.

\param[in] NewTimestamp A #CFE_TIME_SysTime_t data structure containing the desired time to be put into the file’s Standard cFE File Header.

\return Execution status, see \ref CFEReturnCodes, or OSAL status \retval #CFE_STATUS_EXTERNAL_RESOURCE_FAIL \covtest \copybrief CFE_STATUS_EXTERNAL_RESOURCE_FAIL \retval #CFE_SUCCESS \copybrief CFE_SUCCESS

\note This function invokes OSAL API routines and the current implementation may return OSAL error codes to the caller if failure occurs. In a future version of CFE, the status codes will be converted to a value in \ref CFEReturnCodes.