leodos_libcfs::ffi

Function CFE_FS_WriteHeader

pub unsafe extern "C" fn CFE_FS_WriteHeader(
    FileDes: osal_id_t,
    Hdr: *mut CFE_FS_Header_t,
) -> CFE_Status_t

/ /** \brief Write the specified Standard cFE File Header to the specified file

\par Description This API will output the specified #CFE_FS_Header_t variable, with some fields automatically updated, to the specified file as the Standard cFE File Header. This API will automatically populate the following fields in the specified #CFE_FS_Header_t:

  -# \link #CFE_FS_Header_t::ContentType \c ContentType \endlink - Filled with 0x63464531 ('cFE1')
  -# \link #CFE_FS_Header_t::Length \c Length \endlink - Filled with the sizeof(CFE_FS_Header_t)
  -# \link #CFE_FS_Header_t::SpacecraftID \c SpacecraftID \endlink - Filled with the Spacecraft ID
  -# \link #CFE_FS_Header_t::ProcessorID \c ProcessorID \endlink - Filled with the Processor ID
  -# \link #CFE_FS_Header_t::ApplicationID \c ApplicationID \endlink -  Filled with the Application ID
  -# \link #CFE_FS_Header_t::TimeSeconds \c TimeSeconds \endlink - Filled with the Time, in seconds, as obtained
                                                                   by #CFE_TIME_GetTime
  -# \link #CFE_FS_Header_t::TimeSubSeconds \c TimeSubSeconds \endlink - Filled with the Time, subseconds, as
                                                                         obtained by #CFE_TIME_GetTime

\par Assumptions, External Events, and Notes: -# The File has already been successfully opened using #OS_OpenCreate and the caller has a legitimate File Descriptor. -# The \c SubType field has been filled appropriately by the Application. -# The \c Description field has been filled appropriately by the Application. -# File offset behavior: Agnostic on entry since it will move the offset to the start of the file, on success the offset will be at the end of the header, undefined offset behavior for error cases.

\param[in] FileDes File Descriptor obtained from a previous call to #OS_OpenCreate that is associated with the file whose header is to be read.

\param[out] Hdr Pointer to a variable of type #CFE_FS_Header_t @nonnull that will be filled with the contents of the Standard cFE File Header. *Hdr is the contents of the Standard cFE File Header for the specified file.

\return Bytes read or error status from OSAL \retval #CFE_FS_BAD_ARGUMENT \copybrief CFE_FS_BAD_ARGUMENT

\note This function invokes OSAL API routines and the current implementation may return OSAL error codes to the caller if failure occurs. In a future version of CFE, the status codes will be converted to a value in \ref CFEReturnCodes.

\sa #CFE_FS_ReadHeader