leodos_libcfs::ffi

Function CFE_MSG_GenerateChecksum

pub unsafe extern "C" fn CFE_MSG_GenerateChecksum(
    MsgPtr: *mut CFE_MSG_Message_t,
) -> CFE_Status_t

/ /** \brief Calculates and sets the checksum of a message

\par Description This routine calculates the checksum of a message according to an implementation-defined algorithm. Then, it sets the checksum field in the message with the calculated value. The contents and location of this field will depend on the underlying implementation of messages. It may be a checksum, a CRC, or some other algorithm.

\par Assumptions, External Events, and Notes: - If the underlying implementation of messages does not include a checksum field, then this routine will return #CFE_MSG_WRONG_MSG_TYPE

\param[in, out] MsgPtr A pointer to the buffer that contains the message @nonnull.

\return Execution status, see \ref CFEReturnCodes \retval #CFE_SUCCESS \copybrief CFE_SUCCESS \retval #CFE_MSG_BAD_ARGUMENT \copybrief CFE_MSG_BAD_ARGUMENT \retval #CFE_MSG_WRONG_MSG_TYPE \copybrief CFE_MSG_WRONG_MSG_TYPE