leodos_libcfs::ffi

Function CFE_MSG_OriginationAction

pub unsafe extern "C" fn CFE_MSG_OriginationAction(
    MsgPtr: *mut CFE_MSG_Message_t,
    BufferSize: usize,
    IsAcceptable: *mut bool,
) -> CFE_Status_t

/ /** \brief Perform any necessary actions on a newly-created message, prior to sending

\par Description This routine updates and/or appends any necessary fields on a message, is invoked via SB just prior to broadcasting the message. The actions include updating any values that should be computed/updated per message, including: - setting the sequence number - updating the timestamp, if present - computing any error control or checksum fields, if present

      The MSG module implementation determines which header fields meet this criteria
      and how they should be computed.

      The BufferSize parameter indicates the allocation size message of the buffer that
      holds the message (i.e. the message envelope size).  In some implementations, the
      allocated buffer may include extra space in order to append a CRC or digital signature.

\sa CFE_MSG_VerificationAction

\param[inout] MsgPtr A pointer to the buffer that contains the message @nonnull. \param[in] BufferSize The size of the buffer encapsulating the message \param[out] IsAcceptable Output variable to be set, indicates message acceptability @nonnull

\return Execution status, see \ref CFEReturnCodes \retval #CFE_SUCCESS \copybrief CFE_SUCCESS \retval #CFE_MSG_BAD_ARGUMENT \copybrief CFE_MSG_BAD_ARGUMENT