leodos_libcfs::ffi

Function CFE_MSG_VerificationAction

pub unsafe extern "C" fn CFE_MSG_VerificationAction(
    MsgPtr: *const CFE_MSG_Message_t,
    BufferSize: usize,
    IsAcceptable: *mut bool,
) -> CFE_Status_t

/ /** \brief Checks message integrity/acceptability

\par Description This routine validates that any error-control field(s) in the message header matches the expected value.

    The specific function of this API is entirely dependent on the header fields
    and may be a no-op if no error checking is implemented.  In that case, it
    will always output "true".

\note Due to the fact that software bus uses a multicast architecture, this function must not modify the message, as the buffer may be shared among multiple receivers. This should generally be the inverse of CFE_MSG_OriginationAction(), but on the origination side it may update header fields and/or modify the message, on the verification/receive side it must only check those fields, not modify them.

\sa CFE_MSG_OriginationAction

\param[in] MsgPtr Message Pointer @nonnull \param[in] BufferSize The size of the buffer encapsulating the message \param[out] IsAcceptable Output variable to be set, indicates message acceptability @nonnull

\return Execution status, see \ref CFEReturnCodes \retval #CFE_SUCCESS \copybrief CFE_SUCCESS \retval #CFE_MSG_BAD_ARGUMENT \copybrief CFE_MSG_BAD_ARGUMENT