leodos_libcfs::ffi

Function CFE_PSP_GetCFETextSegmentInfo

pub unsafe extern "C" fn CFE_PSP_GetCFETextSegmentInfo(
    PtrToCFESegment: *mut cpuaddr,
    SizeOfCFESegment: *mut uint32,
) -> int32

@brief Returns the location and size of the kernel memory.

This function returns the start and end address of the CFE text segment. It may not be implemented on all architectures.

@param[out] PtrToCFESegment Pointer to the variable that will store the location of the cFE text segment @param[out] SizeOfCFESegment Pointer to the variable that will store the size of the cFE text segment

@return 0 (OS_SUCCESS or CFE_PSP_SUCCESS) on success, -1 (OS_ERROR or CFE_PSP_ERROR) on error