leodos_libcfs::ffi

Function CFE_PSP_Get_Timebase

pub unsafe extern "C" fn CFE_PSP_Get_Timebase(
    Tbu: *mut uint32,
    Tbl: *mut uint32,
)

@brief Sample/Read a monotonic platform clock without normalization

Provides a common interface to system timebase. This routine is in the BSP because it is sometimes implemented in hardware and sometimes taken care of by the RTOS.

This is defined as a free-running, monotonically-increasing tick counter. The epoch is not defined, but typically is the system boot time, and the value increases indefinitely as the system runs. The tick period/rate is also not defined.

Rollover events - where the range of representable values is exceeded - are theoretically possible, but would take many years of continuous uptime to occur (typically hundreds of years, if not thousands). System designers should ensure that the actual tick rate and resulting timebase range is sufficiently large to ensure that rollover is not a concern.

@note This is a “raw” value from the underlying platform with minimal/no conversions or normalization applied. Neither the epoch nor the resolution of this tick counter is specified, and it may vary from platform to platform. Use the CFE_PSP_GetTime() function to sample the timebase and also convert the units into a normalized/more consistent form.

@sa CFE_PSP_GetTime()

@param[out] Tbu Buffer to hold the upper 32 bits of a 64-bit tick counter @param[out] Tbl Buffer to hold the lower 32 bits of a 64-bit tick counter