leodos_libcfs::ffi

Function CFE_ResourceId_FindNext

pub unsafe extern "C" fn CFE_ResourceId_FindNext(
    StartId: CFE_ResourceId_t,
    TableSize: uint32,
    CheckFunc: Option<unsafe extern "C" fn(arg1: CFE_ResourceId_t) -> bool>,
) -> CFE_ResourceId_t

@brief Locate the next resource ID which does not map to an in-use table entry

This begins searching from StartId which should be the most recently issued ID for the resource category. This will then search for the next ID which does not map to a table entry that is in use. That is, it does not alias any valid ID when converted to an array index.

returns an undefined ID value if no open slots are available

@param[in] StartId the last issued ID for the resource category (app, lib, etc). @param[in] TableSize the maximum size of the target table @param[in] CheckFunc a function to check if the given ID is available @returns Next ID value which does not map to a valid entry @retval #CFE_RESOURCEID_UNDEFINED if no open slots or bad arguments.