leodos_libcfs::ffi

Function CFE_SB_AllocateMessageBuffer

pub unsafe extern "C" fn CFE_SB_AllocateMessageBuffer(
    MsgSize: usize,
) -> *mut CFE_SB_Buffer_t

/ /** \brief Get a buffer pointer to use for “zero copy” SB sends.

\par Description This routine can be used to get a pointer to one of the software bus’ internal memory buffers that are used for sending messages. The caller can use this memory buffer to build an SB message, then send it using the CFE_SB_TransmitBuffer() function. This interface avoids an extra copy of the message from the user’s memory buffer to the software bus internal buffer.

\par Assumptions, External Events, and Notes: -# The pointer returned by CFE_SB_AllocateMessageBuffer() is only good for one call to CFE_SB_TransmitBuffer(). -# Once a buffer has been successfully transmitted (as indicated by a successful return from CFE_SB_TransmitBuffer()) the buffer becomes owned by the SB application. It will automatically be freed by SB once all recipients have finished reading it. -# Applications must not de-reference the message pointer (for reading or writing) after the call to CFE_SB_TransmitBuffer(). -# If #CFE_SB_ReleaseMessageBuffer should be used only if a message is not transmitted

\param[in] MsgSize The size of the SB message buffer the caller wants (including the SB message header).

\return A pointer to a memory buffer that message data can be written to for use with CFE_SB_TransmitBuffer().