leodos_libcfs::ffi

Function CFE_SB_CreatePipe

pub unsafe extern "C" fn CFE_SB_CreatePipe(
    PipeIdPtr: *mut CFE_SB_PipeId_t,
    Depth: uint16,
    PipeName: *const c_char,
) -> CFE_Status_t

/ /** \brief Creates a new software bus pipe.

\par Description This routine creates and initializes an input pipe that the calling application can use to receive software bus messages. By default, no messages are routed to the new pipe. So, the application must use #CFE_SB_Subscribe to specify which messages it wants to receive on this pipe.

\par Assumptions, External Events, and Notes: None

\param[out] PipeIdPtr A pointer to a variable of type #CFE_SB_PipeId_t @nonnull, which will be filled in with the pipe ID information by the #CFE_SB_CreatePipe routine. *PipeIdPtr is the identifier for the created pipe.

\param[in] Depth The maximum number of messages that will be allowed on this pipe at one time.

\param[in] PipeName A string @nonnull to be used to identify this pipe in error messages and routing information telemetry. The string must be no longer than #OS_MAX_API_NAME (including terminator). Longer strings will be truncated.

\return Execution status, see \ref CFEReturnCodes \retval #CFE_SUCCESS \copybrief CFE_SUCCESS \retval #CFE_SB_BAD_ARGUMENT \copybrief CFE_SB_BAD_ARGUMENT \retval #CFE_SB_MAX_PIPES_MET \copybrief CFE_SB_MAX_PIPES_MET \retval #CFE_SB_PIPE_CR_ERR \copybrief CFE_SB_PIPE_CR_ERR

\sa #CFE_SB_DeletePipe #CFE_SB_GetPipeOpts #CFE_SB_SetPipeOpts #CFE_SB_GetPipeIdByName