leodos_libcfs::ffi

Function CFE_SB_MessageStringGet

pub unsafe extern "C" fn CFE_SB_MessageStringGet(
    DestStringPtr: *mut c_char,
    SourceStringPtr: *const c_char,
    DefaultString: *const c_char,
    DestMaxSize: usize,
    SourceMaxSize: usize,
) -> int32

/ /** \brief Copies a string out of a software bus message

\par Description Strings within software bus messages have a defined/fixed maximum length, and may not necessarily be null terminated within the message. This presents a possible issue when using the C library functions to copy strings out of a message.

This function should replace use of C library functions such as strcpy/strncpy
when copying strings out of software bus messages to local storage buffers.

Up to [SourceMaxSize] or [DestMaxSize-1] (whichever is smaller) characters will be
copied from the source buffer to the destination buffer, and a NUL termination
character will be written to the destination buffer as the last character.

If the DefaultString pointer is non-NULL, it will be used in place of the source
string if the source is an empty string.  This is typically a string constant that
comes from the platform configuration, allowing default values to be assumed for
fields that are unspecified.

IMPORTANT - the default string, if specified, must be null terminated.  This will
be the case if a string literal is passed in (the typical/expected use case).

If the default is NULL, then only the source string will be copied, and the result
will be an empty string if the source was empty.

If the destination buffer is too small to store the entire string, it will be
truncated, but it will still be null terminated.

\param[out] DestStringPtr Pointer to destination buffer @nonnull \param[in] SourceStringPtr Pointer to source buffer (component of SB message definition) \param[in] DefaultString Default string to use if source is empty \param[in] DestMaxSize Size of destination storage buffer @nonzero \param[in] SourceMaxSize Size of source buffer as defined by the message definition

\return Number of characters copied or error code, see \ref CFEReturnCodes \retval #CFE_SB_BAD_ARGUMENT \copybrief CFE_SB_BAD_ARGUMENT