leodos_libcfs::ffi

Function CFE_SB_MessageStringSet

pub unsafe extern "C" fn CFE_SB_MessageStringSet(
    DestStringPtr: *mut c_char,
    SourceStringPtr: *const c_char,
    DestMaxSize: usize,
    SourceMaxSize: usize,
) -> int32

/ /** \brief Copies a string into a software bus message

\par Description Strings within software bus messages have a defined/fixed maximum length, and may not necessarily be null terminated within the message. This presents a possible issue when using the C library functions to copy strings out of a message.

This performs a very similar function to "strncpy()" except that the sizes
of _both_ buffers are passed in.  Neither buffer is required to be null-terminated,
but copying will stop after the first termination character is encountered.

If the destination buffer is not completely filled by the source data (such as if
the supplied string was shorter than the allotted length) the destination buffer
will be padded with NUL characters up to the size of the buffer, similar to what
strncpy() does.  This ensures that the entire destination buffer is set.

\note If the source string buffer is already guaranteed to be null terminated, then there is no difference between the C library “strncpy()” function and this implementation. It is only necessary to use this when termination of the source buffer is not guaranteed.

\param[out] DestStringPtr Pointer to destination buffer (component of SB message definition) @nonnull \param[in] SourceStringPtr Pointer to source buffer @nonnull \param[in] DestMaxSize Size of destination buffer as defined by the message definition \param[in] SourceMaxSize Size of source buffer

\return Number of characters copied or error code, see \ref CFEReturnCodes \retval #CFE_SB_BAD_ARGUMENT \copybrief CFE_SB_BAD_ARGUMENT