leodos_libcfs::ffi

Function CFE_SB_ReceiveBuffer

pub unsafe extern "C" fn CFE_SB_ReceiveBuffer(
    BufPtr: *mut *mut CFE_SB_Buffer_t,
    PipeId: CFE_SB_PipeId_t,
    TimeOut: int32,
) -> CFE_Status_t

/ /** \brief Receive a message from a software bus pipe

\par Description This routine retrieves the next message from the specified pipe. If the pipe is empty, this routine will block until either a new message comes in or the timeout value is reached.

\par Assumptions, External Events, and Notes: Note - If an error occurs in this API, the *BufPtr value may be NULL or random. Therefore, it is recommended that the return code be tested for CFE_SUCCESS before processing the message.

\param[in, out] BufPtr A pointer to the software bus buffer to receive to @nonnull. Typically a caller declares a ptr of type CFE_SB_Buffer_t (i.e. CFE_SB_Buffer_t *Ptr) then gives the address of that pointer (&Ptr) as this parameter. After a successful receipt of a message, *BufPtr will point to the first byte of the software bus buffer. This should be used as a read-only pointer (in systems with an MMU, writes to this pointer may cause a memory protection fault). The *BufPtr is valid only until the next call to CFE_SB_ReceiveBuffer for the same pipe.

\param[in] PipeId The pipe ID of the pipe containing the message to be obtained.

\param[in] TimeOut The number of milliseconds to wait for a new message if the pipe is empty at the time of the call. This can also be set to #CFE_SB_POLL for a non-blocking receive or #CFE_SB_PEND_FOREVER to wait forever for a message to arrive.

\return Execution status, see \ref CFEReturnCodes \retval #CFE_SUCCESS \copybrief CFE_SUCCESS \retval #CFE_SB_BAD_ARGUMENT \copybrief CFE_SB_BAD_ARGUMENT \retval #CFE_SB_TIME_OUT \copybrief CFE_SB_TIME_OUT \retval #CFE_SB_PIPE_RD_ERR \covtest \copybrief CFE_SB_PIPE_RD_ERR \retval #CFE_SB_NO_MESSAGE \copybrief CFE_SB_NO_MESSAGE