leodos_libcfs::ffi

Function CFE_SB_Subscribe

pub unsafe extern "C" fn CFE_SB_Subscribe(
    MsgId: CFE_SB_MsgId_t,
    PipeId: CFE_SB_PipeId_t,
) -> CFE_Status_t

/ /** \brief Subscribe to a message on the software bus with default parameters

\par Description This routine adds the specified pipe to the destination list for the specified message ID. This is the same as #CFE_SB_SubscribeEx with the Quality field set to #CFE_SB_DEFAULT_QOS and MsgLim set to #CFE_PLATFORM_SB_DEFAULT_MSG_LIMIT (4).

\par Assumptions, External Events, and Notes: Note: As subscriptions are received, the destinations are added to the head of a linked list. During the sending of a message, the list is traversed beginning at the head of the list. Therefore the message will first be sent to the last subscriber. If an application has timing constraints and needs to receive a message in the shortest possible time, the developer may consider holding off its subscription until other applications have subscribed to the message.

\param[in] MsgId The message ID of the message to be subscribed to.

\param[in] PipeId The pipe ID of the pipe the subscribed message should be sent to.

\return Execution status, see \ref CFEReturnCodes \retval #CFE_SUCCESS \copybrief CFE_SUCCESS \retval #CFE_SB_MAX_MSGS_MET \covtest \copybrief CFE_SB_MAX_MSGS_MET \retval #CFE_SB_MAX_DESTS_MET \copybrief CFE_SB_MAX_DESTS_MET \retval #CFE_SB_BAD_ARGUMENT \copybrief CFE_SB_BAD_ARGUMENT \retval #CFE_SB_BUF_ALOC_ERR \covtest \copybrief CFE_SB_BUF_ALOC_ERR

\sa #CFE_SB_SubscribeEx, #CFE_SB_SubscribeLocal, #CFE_SB_Unsubscribe, #CFE_SB_UnsubscribeLocal