leodos_libcfs::ffi

Function CFE_SB_TransmitBuffer

pub unsafe extern "C" fn CFE_SB_TransmitBuffer(
    BufPtr: *mut CFE_SB_Buffer_t,
    IsOrigination: bool,
) -> CFE_Status_t

/ /** \brief Transmit a buffer

\par Description This routine sends a message that has been created directly in an internal SB message buffer by an application (after a call to #CFE_SB_AllocateMessageBuffer). This interface is more complicated than the normal #CFE_SB_TransmitMsg interface, but it avoids an extra copy of the message from the user’s memory buffer to the software bus internal buffer. The “zero copy” interface can be used to improve performance in high-rate, high-volume software bus traffic.

      The IsOrigination parameter should be passed as "true" if the message was
      newly constructed by the sender and is being sent for the first time.  This
      enables the message origination actions as determined by the CFE MSG module,
      which may include (but not limited to):
        - Updating sequence number
        - Updating timestamp
        - Calcualating a CRC, checksum, or other message error control field

      Conversely, when forwarding a message that originated from
      an external entity (e.g. messages passing through CI or SBN), the
      parameter should be passed as "false" to not overwrite existing data.

\par Assumptions, External Events, and Notes: -# A handle returned by #CFE_SB_AllocateMessageBuffer is “consumed” by a successful call to #CFE_SB_TransmitBuffer. -# If this function returns CFE_SUCCESS, this indicates the zero copy handle is now owned by software bus, and is no longer owned by the calling application, and should not be re-used. -# However if this function fails (returns any error status) it does not change the state of the buffer at all, meaning the calling application still owns it. (a failure means the buffer is left in the same state it was before the call). -# Applications should be written as if #CFE_SB_AllocateMessageBuffer is equivalent to a \c malloc() and a successful call to #CFE_SB_TransmitBuffer is equivalent to a \c free(). -# Applications must not de-reference the message pointer (for reading or writing) after a successful call to #CFE_SB_TransmitBuffer. -# This function will increment and apply the internally tracked sequence counter if set to do so.

\param[in] BufPtr A pointer to the buffer to be sent @nonnull. \param[in] IsOrigination Update applicable header field(s) of a newly constructed message

\return Execution status, see \ref CFEReturnCodes \retval #CFE_SUCCESS \copybrief CFE_SUCCESS \retval #CFE_SB_BAD_ARGUMENT \copybrief CFE_SB_BAD_ARGUMENT \retval #CFE_SB_MSG_TOO_BIG \copybrief CFE_SB_MSG_TOO_BIG