leodos_libcfs::ffi

Function CFE_SB_TransmitMsg

pub unsafe extern "C" fn CFE_SB_TransmitMsg(
    MsgPtr: *const CFE_MSG_Message_t,
    IsOrigination: bool,
) -> CFE_Status_t

/ /** \brief Transmit a message

\par Description This routine copies the specified message into a software bus buffer which is then transmitted to all subscribers. The software bus will read the message ID from the message header to determine which pipes should receive the message.

      The IsOrigination parameter should be passed as "true" if the message was
      newly constructed by the sender and is being sent for the first time.  This
      enables the message origination actions as determined by the CFE MSG module,
      which may include (but not limited to):
        - Updating sequence number
        - Updating timestamp
        - Calcualating a CRC, checksum, or other message error control field

      Conversely, when forwarding a message that originated from
      an external entity (e.g. messages passing through CI or SBN), the
      parameter should be passed as "false" to not overwrite existing data.

\par Assumptions, External Events, and Notes: - This routine will not normally wait for the receiver tasks to process the message before returning control to the caller’s task. - However, if a higher priority task is pending and subscribed to this message, that task may get to run before returning control to the caller. - In previous versions of CFE, the boolean parameter referred to the sequence number header of telemetry messages only. This has been extended to apply more generically to any headers, as determined by the CFE MSG implementation.

\param[in] MsgPtr A pointer to the message to be sent @nonnull. This must point to the first byte of the message header. \param[in] IsOrigination Update the headers of the message

\return Execution status, see \ref CFEReturnCodes \retval #CFE_SUCCESS \copybrief CFE_SUCCESS \retval #CFE_SB_BAD_ARGUMENT \copybrief CFE_SB_BAD_ARGUMENT \retval #CFE_SB_MSG_TOO_BIG \copybrief CFE_SB_MSG_TOO_BIG \retval #CFE_SB_BUF_ALOC_ERR \covtest \copybrief CFE_SB_BUF_ALOC_ERR