leodos_libcfs::ffi

Function CFE_TBL_GetAddress

pub unsafe extern "C" fn CFE_TBL_GetAddress(
    TblPtr: *mut *mut c_void,
    TblHandle: CFE_TBL_Handle_t,
) -> CFE_Status_t

/ /** \brief Obtain the current address of the contents of the specified table

\par Description When a table has been created and initialized, it is available to any application that can identify it with its unique handle. In order to view the data contained in the table, an application must call this function or #CFE_TBL_GetAddresses.

\par Assumptions, External Events, and Notes: -# This call can be a blocking call when the table is not double buffered and is shared with another application of lower priority that just happens to be in the middle of a table update of the specific table. If this occurs, the application performing the table update will automatically have its priority elevated in order to release the resource as soon as possible. -# An application must always release the returned table address using the #CFE_TBL_ReleaseAddress or #CFE_TBL_ReleaseAddresses function prior to either a #CFE_TBL_Update call or any blocking call (e.g. - pending on software bus message, etc). Table updates cannot occur while table addresses have not been released. -# #CFE_TBL_ERR_NEVER_LOADED will be returned if the table has never been loaded (either from file or from a block of memory), but the function will still return a valid table pointer to a table with all zero content. This pointer must be released with the #CFE_TBL_ReleaseAddress API before the table can be loaded with data.

\param[out] TblPtr The address of a pointer @nonnull that will be loaded with the address of the first byte of the table. This pointer can then be typecast by the calling application to the appropriate table data structure. *TblPtr is the address of the first byte of data associated with the specified table.

\param[in] TblHandle Handle, previously obtained from #CFE_TBL_Register or #CFE_TBL_Share, that identifies the Table whose address is to be returned.

\return Execution status, see \ref CFEReturnCodes \retval #CFE_SUCCESS \copybrief CFE_SUCCESS \retval #CFE_TBL_INFO_UPDATED \copybrief CFE_TBL_INFO_UPDATED \retval #CFE_ES_ERR_RESOURCEID_NOT_VALID \copybrief CFE_ES_ERR_RESOURCEID_NOT_VALID \retval #CFE_TBL_ERR_NO_ACCESS \copybrief CFE_TBL_ERR_NO_ACCESS \retval #CFE_TBL_ERR_INVALID_HANDLE \copybrief CFE_TBL_ERR_INVALID_HANDLE \retval #CFE_TBL_ERR_UNREGISTERED \copybrief CFE_TBL_ERR_UNREGISTERED \retval #CFE_TBL_ERR_NEVER_LOADED \copybrief CFE_TBL_ERR_NEVER_LOADED \retval #CFE_TBL_BAD_ARGUMENT \copybrief CFE_TBL_BAD_ARGUMENT

\sa #CFE_TBL_ReleaseAddress, #CFE_TBL_GetAddresses, #CFE_TBL_ReleaseAddresses