leodos_libcfs::ffi

Function CFE_TBL_Load

pub unsafe extern "C" fn CFE_TBL_Load(
    TblHandle: CFE_TBL_Handle_t,
    SrcType: CFE_TBL_SrcEnum_t,
    SrcDataPtr: *const c_void,
) -> CFE_Status_t

/ /** \brief Load a specified table with data from specified source

\par Description Once an application has created a table (#CFE_TBL_Register), it must provide the values that initialize the contents of that table. The application accomplishes this with one of two different TBL API calls. This function call initializes the table with values that are held in a data structure.

\par Assumptions, External Events, and Notes: This function call can block. Therefore, interrupt service routines should NOT initialize their own tables. An application should initialize any table(s) prior to providing the handle(s) to the interrupt service routine.

\param[in] TblHandle Handle, previously obtained from #CFE_TBL_Register or #CFE_TBL_Share, that identifies the Table to be loaded.

\param[in] SrcType Flag indicating the nature of the given \c SrcDataPtr below. This value can be any one of the following: \arg #CFE_TBL_SRC_FILE - \copybrief CFE_TBL_SRC_FILE \arg #CFE_TBL_SRC_ADDRESS - \copybrief CFE_TBL_SRC_ADDRESS

\param[in] SrcDataPtr Pointer @nonnull to either a character string specifying a filename or a memory address of a block of binary data to be loaded into a table or, if the table was registered with the #CFE_TBL_OPT_USR_DEF_ADDR option, the address of the active table buffer.

\return Execution status, see \ref CFEReturnCodes \retval #CFE_SUCCESS \copybrief CFE_SUCCESS \retval #CFE_ES_ERR_RESOURCEID_NOT_VALID \copybrief CFE_ES_ERR_RESOURCEID_NOT_VALID \retval #CFE_TBL_ERR_NO_ACCESS \copybrief CFE_TBL_ERR_NO_ACCESS \retval #CFE_TBL_ERR_INVALID_HANDLE \copybrief CFE_TBL_ERR_INVALID_HANDLE \retval #CFE_TBL_ERR_DUMP_ONLY \copybrief CFE_TBL_ERR_DUMP_ONLY \retval #CFE_TBL_ERR_ILLEGAL_SRC_TYPE \copybrief CFE_TBL_ERR_ILLEGAL_SRC_TYPE \retval #CFE_TBL_ERR_LOAD_IN_PROGRESS \copybrief CFE_TBL_ERR_LOAD_IN_PROGRESS \retval #CFE_TBL_ERR_LOAD_INCOMPLETE \copybrief CFE_TBL_ERR_LOAD_INCOMPLETE \retval #CFE_TBL_ERR_NO_BUFFER_AVAIL \copybrief CFE_TBL_ERR_NO_BUFFER_AVAIL \retval #CFE_TBL_ERR_ACCESS \copybrief CFE_TBL_ERR_ACCESS \retval #CFE_TBL_ERR_FILE_TOO_LARGE \copybrief CFE_TBL_ERR_FILE_TOO_LARGE \retval #CFE_TBL_ERR_BAD_CONTENT_ID \copybrief CFE_TBL_ERR_BAD_CONTENT_ID \retval #CFE_TBL_ERR_BAD_SUBTYPE_ID \copybrief CFE_TBL_ERR_BAD_SUBTYPE_ID \retval #CFE_TBL_ERR_NO_STD_HEADER \copybrief CFE_TBL_ERR_NO_STD_HEADER \retval #CFE_TBL_ERR_NO_TBL_HEADER \copybrief CFE_TBL_ERR_NO_TBL_HEADER \retval #CFE_TBL_ERR_PARTIAL_LOAD \copybrief CFE_TBL_ERR_PARTIAL_LOAD \retval #CFE_TBL_BAD_ARGUMENT \copybrief CFE_TBL_BAD_ARGUMENT

\sa #CFE_TBL_Update, #CFE_TBL_Validate, #CFE_TBL_Manage