leodos_libcfs::ffi

Function CFE_TBL_NotifyByMessage

pub unsafe extern "C" fn CFE_TBL_NotifyByMessage(
    TblHandle: CFE_TBL_Handle_t,
    MsgId: CFE_SB_MsgId_t,
    CommandCode: CFE_MSG_FcnCode_t,
    Parameter: uint32,
) -> CFE_Status_t

/ /** \brief Instruct cFE Table Services to notify Application via message when table requires management

\par Description This API instructs Table Services to send a message to the calling Application whenever the specified table requires management by the application. This feature allows applications to avoid polling table services via the #CFE_TBL_Manage call to determine whether a table requires updates, validation, etc. This API should be called following the #CFE_TBL_Register API whenever the owning application requires this feature.

\par Assumptions, External Events, and Notes: - Only the application that owns the table is allowed to register a notification message - Recommend \b NOT using the ground command MID which typically impacts command counters. The typical approach is to use a unique MID for inter-task communications similar to how schedulers typically trigger application housekeeping messages.

\param[in] TblHandle Handle of Table with which the message should be associated.

\param[in] MsgId Message ID to be used in notification message sent by Table Services.

\param[in] CommandCode Command Code value to be placed in secondary header of message sent by Table Services.

\param[in] Parameter Application defined value to be passed as a parameter in the message sent by Table Services. Suggested use includes an application’s table index that allows the same MsgId and Command Code to be used for all table management notifications.

\return Execution status, see \ref CFEReturnCodes \retval #CFE_SUCCESS \copybrief CFE_SUCCESS \retval #CFE_ES_ERR_RESOURCEID_NOT_VALID \copybrief CFE_ES_ERR_RESOURCEID_NOT_VALID \retval #CFE_TBL_ERR_NO_ACCESS \copybrief CFE_TBL_ERR_NO_ACCESS \retval #CFE_TBL_ERR_INVALID_HANDLE \copybrief CFE_TBL_ERR_INVALID_HANDLE

\sa #CFE_TBL_Register