leodos_libcfs::ffi

Function CFE_TBL_Register

pub unsafe extern "C" fn CFE_TBL_Register(
    TblHandlePtr: *mut CFE_TBL_Handle_t,
    Name: *const c_char,
    Size: usize,
    TblOptionFlags: uint16,
    TblValidationFuncPtr: CFE_TBL_CallbackFuncPtr_t,
) -> CFE_Status_t

/ /** \brief Register a table with cFE to obtain Table Management Services

\par Description When an application is created and initialized, it is responsible for creating its table images via the TBL API. The application must inform the Table Service of the table name, table size and selection of optional table features.

\par Assumptions, External Events, and Notes: Note: This function call can block. Therefore, interrupt service routines should NOT create their own tables. An application should create any table(s) and provide the handle(s) to the interrupt service routine.

\param[out] TblHandlePtr a pointer to a #CFE_TBL_Handle_t type variable @nonnull that will be assigned the table’s handle. The table handle is required for other API calls when accessing the data contained in the table. *TblHandlePtr is the handle used to identify table to cFE when performing Table operations. This value is returned at address specified by TblHandlePtr.

\param[in] Name The raw table name. This name will be combined with the name of the application to produce a name of the form “AppName.RawTableName”. This application specific name will be used in commands for modifying or viewing the contents of the table.

\param[in] Size The size, in bytes, of the table to be created @nonzero. This is the size that will be allocated as a shared memory resource between the Table Management Service and the calling application.

\param[in] TblOptionFlags Flag bits indicating selected options for table. A bitwise OR of the following option flags: \arg #CFE_TBL_OPT_DEFAULT - The default setting for table options is a combination of #CFE_TBL_OPT_SNGL_BUFFER and #CFE_TBL_OPT_LOAD_DUMP. See below for a description of these two options. This option is mutually exclusive with the #CFE_TBL_OPT_DBL_BUFFER, #CFE_TBL_OPT_DUMP_ONLY and #CFE_TBL_OPT_USR_DEF_ADDR options. \arg #CFE_TBL_OPT_SNGL_BUFFER - When this option is selected, the table will use a shared session table for performing table modifications and a memory copy from the session table to the “active” table buffer will occur when the table is updated. This is the preferred option since it will minimize memory usage. This option is mutually exclusive with the #CFE_TBL_OPT_DBL_BUFFER option \arg #CFE_TBL_OPT_DBL_BUFFER - When this option is selected, two instances of the table are created. One is considered the “active” table and the other the “inactive” table. Whenever table modifications occur, they do not require the use of a common session table. Modifications occur in the “inactive” buffer. Then, when it is time to update the table, the pointer to the “active” table is changed to point to the “inactive” buffer thus making it the new “active” buffer. This feature is most useful for time critical applications (ie - interrupt service routines, etc). This option is mutually exclusive with the #CFE_TBL_OPT_SNGL_BUFFER and #CFE_TBL_OPT_DEFAULT option. \arg #CFE_TBL_OPT_LOAD_DUMP - When this option is selected, the Table Service is allowed to perform all operations on the specified table. This option is mutually exclusive with the #CFE_TBL_OPT_DUMP_ONLY option. \arg #CFE_TBL_OPT_DUMP_ONLY - When this option is selected, the Table Service will not perform table loads to this table. This does not prevent, however, a task from writing to the table via an address obtained with the #CFE_TBL_GetAddress API function. This option is mutually exclusive with the #CFE_TBL_OPT_LOAD_DUMP and #CFE_TBL_OPT_DEFAULT options. If the Application wishes to specify their own block of memory as the Dump Only table, they need to also include the #CFE_TBL_OPT_USR_DEF_ADDR option explained below. \arg #CFE_TBL_OPT_NOT_USR_DEF - When this option is selected, Table Services allocates memory for the table and, in the case of a double buffered table, it allocates the same amount of memory again for the second buffer. This option is mutually exclusive with the #CFE_TBL_OPT_USR_DEF_ADDR option. \arg #CFE_TBL_OPT_USR_DEF_ADDR- When this option is selected, the Table Service will not allocate memory for the table. Table Services will require the Application to identify the location of the active table buffer via the #CFE_TBL_Load function. This option implies the #CFE_TBL_OPT_DUMP_ONLY and the #CFE_TBL_OPT_SNGL_BUFFER options and is mutually exclusive of the #CFE_TBL_OPT_DBL_BUFFER option. \arg #CFE_TBL_OPT_CRITICAL- When this option is selected, the Table Service will automatically allocate space in the Critical Data Store (CDS) for the table and ensure that the contents in the CDS are the same as the contents of the currently active buffer for the table. This option is mutually exclusive of the #CFE_TBL_OPT_USR_DEF_ADDR and #CFE_TBL_OPT_DUMP_ONLY options. It should also be noted that the use of this option with double buffered tables will prevent the update of the double buffered table from being quick and it could be blocked. Therefore, critical tables should not be updated by Interrupt Service Routines.

\param[in] TblValidationFuncPtr is a pointer to a function that will be executed in the context of the Table Management Service when the contents of a table need to be validated. If set to NULL, then the Table Management Service will assume any data is valid. If the value is not NULL, it must be a pointer to a function with the following prototype:
int32 CallbackFunc(void *TblPtr);
where
TblPtr will be a pointer to the table data that is to be verified. When the function returns #CFE_SUCCESS, the data is considered valid and ready for a commit. When the function returns a negative value, the data is considered invalid and an Event Message will be issued containing the returned value. If the function should return a positive number, the table is considered invalid and the return code is considered invalid. Validation functions \b must return either #CFE_SUCCESS or a negative number (whose value is at the developer’s discretion). The validation function will be executed in the Application’s context so that Event Messages describing the validation failure are possible from within the function.

\return Execution status, see \ref CFEReturnCodes \retval #CFE_SUCCESS \copybrief CFE_SUCCESS \retval #CFE_TBL_INFO_RECOVERED_TBL \copybrief CFE_TBL_INFO_RECOVERED_TBL \retval #CFE_TBL_ERR_DUPLICATE_DIFF_SIZE \copybrief CFE_TBL_ERR_DUPLICATE_DIFF_SIZE \retval #CFE_TBL_ERR_DUPLICATE_NOT_OWNED \copybrief CFE_TBL_ERR_DUPLICATE_NOT_OWNED \retval #CFE_TBL_ERR_REGISTRY_FULL \copybrief CFE_TBL_ERR_REGISTRY_FULL \retval #CFE_TBL_ERR_HANDLES_FULL \copybrief CFE_TBL_ERR_HANDLES_FULL \retval #CFE_TBL_ERR_INVALID_SIZE \copybrief CFE_TBL_ERR_INVALID_SIZE \retval #CFE_TBL_ERR_INVALID_NAME \copybrief CFE_TBL_ERR_INVALID_NAME \retval #CFE_ES_ERR_RESOURCEID_NOT_VALID \copybrief CFE_ES_ERR_RESOURCEID_NOT_VALID \retval #CFE_TBL_BAD_ARGUMENT \copybrief CFE_TBL_BAD_ARGUMENT \retval #CFE_TBL_ERR_INVALID_OPTIONS \copybrief CFE_TBL_ERR_INVALID_OPTIONS \retval #CFE_TBL_WARN_DUPLICATE \copybrief CFE_TBL_WARN_DUPLICATE \retval #CFE_TBL_WARN_NOT_CRITICAL \copybrief CFE_TBL_WARN_NOT_CRITICAL

\sa #CFE_TBL_Unregister, #CFE_TBL_Share