leodos_libcfs::ffi

Function CFE_TBL_ReleaseAddresses

pub unsafe extern "C" fn CFE_TBL_ReleaseAddresses(
    NumTables: uint16,
    TblHandles: *const CFE_TBL_Handle_t,
) -> CFE_Status_t

/ /** \brief Release the addresses of an array of specified tables

\par Description Each application is \b required to release a table address obtained through the #CFE_TBL_GetAddress function.

\par Assumptions, External Events, and Notes: An application must always release the returned table address using the #CFE_TBL_ReleaseAddress function prior to either a #CFE_TBL_Update call or any blocking call (e.g. - pending on software bus message, etc). Table updates cannot occur while table addresses have not been released.

\param[in] NumTables Size of TblHandles array.

\param[in] TblHandles Array of Table Handles @nonnull, previously obtained from #CFE_TBL_Register or #CFE_TBL_Share, of those tables whose start addresses are to be released.

\return Execution status, see \ref CFEReturnCodes \retval #CFE_SUCCESS \copybrief CFE_SUCCESS \retval #CFE_TBL_INFO_UPDATED \copybrief CFE_TBL_INFO_UPDATED \retval #CFE_ES_ERR_RESOURCEID_NOT_VALID \copybrief CFE_ES_ERR_RESOURCEID_NOT_VALID \retval #CFE_TBL_ERR_NO_ACCESS \copybrief CFE_TBL_ERR_NO_ACCESS \retval #CFE_TBL_ERR_INVALID_HANDLE \copybrief CFE_TBL_ERR_INVALID_HANDLE \retval #CFE_TBL_ERR_NEVER_LOADED \copybrief CFE_TBL_ERR_NEVER_LOADED \retval #CFE_TBL_BAD_ARGUMENT \copybrief CFE_TBL_BAD_ARGUMENT

\sa #CFE_TBL_GetAddress, #CFE_TBL_ReleaseAddress, #CFE_TBL_GetAddresses