leodos_libcfs::ffi

Function CFE_TIME_ExternalTime

pub unsafe extern "C" fn CFE_TIME_ExternalTime(
    NewTime: CFE_TIME_SysTime_t,
)

/ /** \brief Provide the time from an external source that measures time relative to a known epoch.

\par Description This routine provides a method to provide cFE TIME with current time data acquired from an external source. There is a presumption that this function will be called at the appropriate time (relative to the tone) such that this call may be used by cFE TIME as the signal to generate the “time at the tone” data command. The “time at the tone” data command must arrive within the configuration specified window for tone signal and data packet verification.

Internally, cFE TIME will calculate a new STCF as the difference between this new time value and the spacecraft MET value at the tone. This allows cFE TIME to always calculate time as the sum of MET and STCF. The value of STCF will change only as much as the drift factor between spacecraft MET and the external time source.

\par Assumptions, External Events, and Notes: - This routine is included in the API only when 3 specific configuration parameters are set to true. The first is #CFE_PLATFORM_TIME_CFG_SERVER which defines this instantiation of cFE TIME as a time server (not a client). The second required configuration parameter is #CFE_PLATFORM_TIME_CFG_SOURCE which enables time source selection commands to the cFE TIME task, and further enables configuration definitions for the selected type of external time data. The third configuration parameter required for this routine is #CFE_PLATFORM_TIME_CFG_SRC_TIME, which indicates that the external time data consists of a time value relative to a known epoch.

\param[in] NewTime The MET value at the next (or previous) 1 Hz tone signal.

\sa #CFE_TIME_ExternalTone, #CFE_TIME_ExternalMET, #CFE_TIME_ExternalGPS