leodos_libcfs::ffi

Function CFE_TIME_GetLeapSeconds

pub unsafe extern "C" fn CFE_TIME_GetLeapSeconds() -> int16

/ /** \brief Get the current value of the leap seconds counter.

\par Description This routine returns the current value of the leap seconds counter. This is the delta seconds between international atomic time (TAI) and universal coordinated time (UTC). There is no API provided to set or adjust leap seconds or SCTF, those actions should be done by command only. This API is provided for applications to be able to include leap seconds in their data products to aid in time correlation during downstream science data processing. Note that some mission operations teams do not maintain the leap seconds count, preferring to adjust the STCF instead. Users of this function should check with their mission ops team to see how they are planning to handle leap seconds.

\par Assumptions, External Events, and Notes: None

\returns The current spacecraft leap seconds.

\sa #CFE_TIME_GetSTCF, #CFE_TIME_GetClockState, #CFE_TIME_GetClockInfo