leodos_libcfs::ffi

Function CFE_TIME_GetTime

pub unsafe extern "C" fn CFE_TIME_GetTime() -> CFE_TIME_SysTime_t

/ /** \brief Get the current spacecraft time

\par Description This routine returns the current spacecraft time, which is the amount of time elapsed since the epoch as set in mission configuration. The time returned is either TAI (no leap seconds) or UTC (including leap seconds). This choice is made in the mission configuration file by defining either #CFE_MISSION_TIME_CFG_DEFAULT_TAI or #CFE_MISSION_TIME_CFG_DEFAULT_UTC as true at compile time. To maintain re-usability across missions, most applications should be using this function rather than the specific routines for getting UTC/TAI directly.

\par Assumptions, External Events, and Notes: None

\return The current spacecraft time in default format

\sa #CFE_TIME_GetTAI, #CFE_TIME_GetUTC, #CFE_TIME_GetMET, #CFE_TIME_GetMETseconds, #CFE_TIME_GetMETsubsecs