leodos_libcfs::ffi

Function CFE_TIME_Print

pub unsafe extern "C" fn CFE_TIME_Print(
    PrintBuffer: *mut c_char,
    TimeToPrint: CFE_TIME_SysTime_t,
)

/ /** \brief Print a time value as a string

\par Description This routine prints the specified time to the specified string buffer in the following format:

\c yyyy-ddd-hh:mm:ss.xxxxx\0

where: - \c yyyy = year - \c ddd = Julian day of the year - \c hh = hour of the day (0 to 23) - \c mm = minute (0 to 59) - \c ss = second (0 to 59) - \c xxxxx = subsecond formatted as a decimal fraction (1/4 second = 0.25000) - \c \0 = trailing null

\par Assumptions, External Events, and Notes: - The value of the time argument is simply added to the configuration definitions for the ground epoch and converted into a fixed length string in the buffer provided by the caller. - A loss of data during the string conversion will occur if the computed year exceeds 9999. However, a year that large would require an unrealistic definition for the ground epoch since the maximum amount of time represented by a CFE_TIME_SysTime structure is approximately 136 years.

\param[out] PrintBuffer Pointer to a character array @nonnull of at least #CFE_TIME_PRINTED_STRING_SIZE characters in length. *PrintBuffer is the time as a character string as described above.

\param[in] TimeToPrint The time to print into the character array.