leodos_libcfs::ffi

Function OS_CondVarCreate

pub unsafe extern "C" fn OS_CondVarCreate(
    var_id: *mut osal_id_t,
    var_name: *const c_char,
    options: uint32,
) -> int32

@brief Creates a condition variable resource

A condition variable adds a more sophisticated synchronization option for mutexes, such that it can operate on arbitrary user-defined conditions rather than simply a counter or boolean (as in the case of simple semaphores).

Creating a condition variable resource in OSAL will in turn create both a basic mutex as well as a synchronization overlay. The underlying mutex is similar to the mutex functionality provided by the OSAL mutex subsystem, and can be locked and unlocked normally.

This mutex is intended to protect access to any arbitrary user-defined data object that serves as the condition being tested.

A task that needs a particular state of the object should follow this general flow:

Lock the underlying mutex

Test for the condition being waited for (a user-defined check on user-defined data)

If condition IS NOT met, then call OS_CondVarWait() to wait, then repeat test

If condition IS met, then unlock the underlying mutex and continue

A task that changes the state of the object should follow this general flow:

Lock the underlying mutex

Change the state as necessary

Call either OS_CondVarSignal() or OS_CondVarBroadcast()

Unlock the underlying mutex

@param[out] var_id will be set to the non-zero ID of the newly-created resource @nonnull @param[in] var_name the name of the new resource to create @nonnull @param[in] options reserved for future use. Should be passed as 0.

@return Execution status, see @ref OSReturnCodes @retval #OS_SUCCESS @copybrief OS_SUCCESS @retval #OS_INVALID_POINTER if var_id or var_name are NULL @retval #OS_ERR_NAME_TOO_LONG name length including null terminator greater than #OS_MAX_API_NAME @retval #OS_ERR_NO_FREE_IDS if there are no more free condition variable Ids @retval #OS_ERR_NAME_TAKEN if there is already a condition variable with the same name