leodos_libcfs::ffi

Function OS_ConvertToArrayIndex

pub unsafe extern "C" fn OS_ConvertToArrayIndex(
    object_id: osal_id_t,
    ArrayIndex: *mut osal_index_t,
) -> int32

@brief Converts an abstract ID into a number suitable for use as an array index.

This will return a unique zero-based integer number in the range of [0,MAX) for any valid object ID. This may be used by application code as an array index for indexing into local tables.

@note This does NOT verify the validity of the ID, that is left to the caller. This is only the conversion logic.

This routine accepts any object type, and returns a value based on the maximum number of objects for that type. This is equivalent to invoking OS_ObjectIdToArrayIndex() with the idtype set to OS_OBJECT_TYPE_UNDEFINED.

@sa OS_ObjectIdToArrayIndex

@param[in] object_id The object ID to operate on @param[out] *ArrayIndex The Index to return @nonnull

@returns Execution status, see @ref OSReturnCodes @retval #OS_SUCCESS @copybrief OS_SUCCESS @retval #OS_ERR_INVALID_ID if the object_id argument is not valid @retval #OS_INVALID_POINTER if the ArrayIndex is NULL