leodos_libcfs::ffi

Function OS_CountSemCreate

pub unsafe extern "C" fn OS_CountSemCreate(
    sem_id: *mut osal_id_t,
    sem_name: *const c_char,
    sem_initial_value: uint32,
    options: uint32,
) -> int32

@brief Creates a counting semaphore

Creates a counting semaphore with initial value specified by sem_initial_value and name specified by sem_name. sem_id will be returned to the caller.

@note Underlying RTOS implementations may or may not impose a specific upper limit to the value of a counting semaphore. If the OS has a specific limit and the sem_initial_value exceeds this limit, then #OS_INVALID_SEM_VALUE is returned. On other implementations, any 32-bit integer value may be acceptable. For maximum portability, it is recommended to keep counting semaphore values within the range of a “short int” (i.e. between 0 and 32767). Many platforms do accept larger values, but may not be guaranteed.

@param[out] sem_id will be set to the non-zero ID of the newly-created resource @nonnull @param[in] sem_name the name of the new resource to create @nonnull @param[in] sem_initial_value the initial value of the counting semaphore @param[in] options Reserved for future use, should be passed as 0.

@return Execution status, see @ref OSReturnCodes @retval #OS_SUCCESS @copybrief OS_SUCCESS @retval #OS_INVALID_POINTER if sen name or sem_id are NULL @retval #OS_ERR_NAME_TOO_LONG name length including null terminator greater than #OS_MAX_API_NAME @retval #OS_ERR_NO_FREE_IDS if all of the semaphore ids are taken @retval #OS_ERR_NAME_TAKEN if this is already the name of a counting semaphore @retval #OS_INVALID_SEM_VALUE if the semaphore value is too high @covtest @retval #OS_SEM_FAILURE if an unspecified implementation error occurs @covtest