leodos_libcfs::ffi

Function OS_FileSysAddFixedMap

pub unsafe extern "C" fn OS_FileSysAddFixedMap(
    filesys_id: *mut osal_id_t,
    phys_path: *const c_char,
    virt_path: *const c_char,
) -> int32

@brief Create a fixed mapping between an existing directory and a virtual OSAL mount point.

This mimics the behavior of a “FS_BASED” entry in the VolumeTable but is registered at runtime. It is intended to be called by the PSP/BSP prior to starting the application.

@note OSAL virtual mount points are required to be a single, non-empty top-level directory name. Virtual path names always follow the form /<virt_mount_point>/<relative_path>/<file>. Only the relative path may be omitted/empty (i.e. /<virt_mount_point>/<file>) but the virtual mount point must be present and not an empty string. In particular this means it is not possible to directly refer to files in the “root” of the native file system from OSAL. However it is possible to create a virtual map to the root, such as by calling:

 OS_FileSysAddFixedMap(&fs_id, "/", "/root");

@param[out] filesys_id A buffer to store the ID of the file system mapping @nonnull @param[in] phys_path The native system directory (an existing mount point) @nonnull @param[in] virt_path The virtual mount point of this filesystem @nonnull

@return Execution status, see @ref OSReturnCodes @retval #OS_SUCCESS @copybrief OS_SUCCESS @retval #OS_FS_ERR_PATH_TOO_LONG if the overall phys_path is too long @retval #OS_ERR_NAME_TOO_LONG if the phys_path basename (filesystem name) is too long @retval #OS_INVALID_POINTER if any argument is NULL