leodos_libcfs::ffi

Function OS_GetResourceName

pub unsafe extern "C" fn OS_GetResourceName(
    object_id: osal_id_t,
    buffer: *mut c_char,
    buffer_size: usize,
) -> int32

@brief Obtain the name of an object given an arbitrary object ID

All OSAL resources generally have a name associated with them. This allows application code to retrieve the name of any valid OSAL object ID.

@param[in] object_id The object ID to operate on @param[out] buffer Buffer in which to store the name @nonnull @param[in] buffer_size Size of the output storage buffer @nonzero

@returns Execution status, see @ref OSReturnCodes @retval #OS_SUCCESS @copybrief OS_SUCCESS @retval #OS_ERR_INVALID_ID if the passed-in ID is not a valid OSAL ID @retval #OS_INVALID_POINTER if the passed-in buffer is invalid @retval #OS_ERR_NAME_TOO_LONG if the name will not fit in the buffer provided