leodos_libcfs::ffi

Function OS_ModuleLoad

pub unsafe extern "C" fn OS_ModuleLoad(
    module_id: *mut osal_id_t,
    module_name: *const c_char,
    filename: *const c_char,
    flags: uint32,
) -> int32

@brief Loads an object file

Loads an object file into the running operating system

The “flags” parameter may influence how the loaded module symbols are made available for use in the application. See #OS_MODULE_FLAG_LOCAL_SYMBOLS and #OS_MODULE_FLAG_GLOBAL_SYMBOLS for descriptions.

@param[out] module_id Non-zero OSAL ID corresponding to the loaded module @param[in] module_name Name of module @nonnull @param[in] filename File containing the object code to load @nonnull @param[in] flags Options for the loaded module

@return Execution status, see @ref OSReturnCodes @retval #OS_SUCCESS @copybrief OS_SUCCESS @retval #OS_INVALID_POINTER if one of the parameters is NULL @retval #OS_ERR_NO_FREE_IDS if the module table is full @retval #OS_ERR_NAME_TAKEN if the name is in use @retval #OS_ERR_NAME_TOO_LONG if the module_name is too long @retval #OS_FS_ERR_PATH_INVALID if the filename argument is not valid @retval #OS_ERROR if an other/unspecified error occurs @covtest