leodos_libcfs::ffi

Function OS_ObjectIdToArrayIndex

pub unsafe extern "C" fn OS_ObjectIdToArrayIndex(
    idtype: osal_objtype_t,
    object_id: osal_id_t,
    ArrayIndex: *mut osal_index_t,
) -> int32

@brief Converts an abstract ID into a number suitable for use as an array index.

This will return a unique zero-based integer number in the range of [0,MAX) for any valid object ID. This may be used by application code as an array index for indexing into local tables.

This routine operates on a specific object type, and returns a value based on the maximum number of objects for that type.

If the idtype is passed as #OS_OBJECT_TYPE_UNDEFINED, then object type verification is skipped and any object ID will be accepted and converted to an index. In this mode, the range of the output depends on the actual passed-in object type.

If the idtype is passed as any other value, the passed-in ID value is first confirmed to be the correct type. This check will guarantee that the output is within an expected range; for instance, if the type is passed as #OS_OBJECT_TYPE_OS_TASK, then the output index is guaranteed to be between 0 and #OS_MAX_TASKS-1 after successful conversion.

@param[in] idtype The object type to convert @param[in] object_id The object ID to operate on @param[out] *ArrayIndex The Index to return @nonnull

@returns Execution status, see @ref OSReturnCodes @retval #OS_SUCCESS @copybrief OS_SUCCESS @retval #OS_ERR_INVALID_ID if the object_id argument is not valid @retval #OS_INVALID_POINTER if the ArrayIndex is NULL