leodos_libcfs::ffi

Function OS_QueueCreate

pub unsafe extern "C" fn OS_QueueCreate(
    queue_id: *mut osal_id_t,
    queue_name: *const c_char,
    queue_depth: osal_blockcount_t,
    data_size: usize,
    flags: uint32,
) -> int32

@brief Create a message queue

This is the function used to create a queue in the operating system. Depending on the underlying operating system, the memory for the queue will be allocated automatically or allocated by the code that sets up the queue. Queue names must be unique; if the name already exists this function fails. Names cannot be NULL.

@param[out] queue_id will be set to the non-zero ID of the newly-created resource @nonnull @param[in] queue_name the name of the new resource to create @nonnull @param[in] queue_depth the maximum depth of the queue @param[in] data_size the size of each entry in the queue @nonzero @param[in] flags options for the queue (reserved for future use, pass as 0)

@return Execution status, see @ref OSReturnCodes @retval #OS_SUCCESS @copybrief OS_SUCCESS @retval #OS_INVALID_POINTER if a pointer passed in is NULL @retval #OS_ERR_NAME_TOO_LONG name length including null terminator greater than #OS_MAX_API_NAME @retval #OS_ERR_NO_FREE_IDS if there are already the max queues created @retval #OS_ERR_NAME_TAKEN if the name is already being used on another queue @retval #OS_ERR_INVALID_SIZE if data_size is 0 @retval #OS_QUEUE_INVALID_SIZE if the queue depth exceeds the limit @retval #OS_ERROR if the OS create call fails