leodos_libcfs::ffi

Function OS_SelectMultiple

pub unsafe extern "C" fn OS_SelectMultiple(
    ReadSet: *mut OS_FdSet,
    WriteSet: *mut OS_FdSet,
    msecs: int32,
) -> int32

@brief Wait for events across multiple file handles

Wait for any of the given sets of IDs to become readable or writable

This function will block until any of the following occurs:

At least one OSAL ID in the ReadSet is readable

At least one OSAL ID in the WriteSet is writable

The timeout has elapsed

The sets are input/output parameters. On entry, these indicate the file handle(s) to wait for. On exit, these are set to the actual file handle(s) that have activity.

If the timeout occurs this returns an error code and all output sets should be empty.

The timeout is expressed in milliseconds, relative to the time that the API was invoked. Use OS_SelectMultipleAbs() for higher timing precision.

@note This does not lock or otherwise protect the file handles in the given sets. If a filehandle supplied via one of the FdSet arguments is closed or modified by another while this function is in progress, the results are undefined. Because of this limitation, it is recommended to use OS_SelectSingle() whenever possible.

@param[in,out] ReadSet Set of handles to check/wait to become readable @param[in,out] WriteSet Set of handles to check/wait to become writable @param[in] msecs Indicates the timeout. Positive values will wait up to that many milliseconds. Zero will not wait (poll). Negative values will wait forever (pend)

@sa OS_SelectMultipleAbs()

@return Execution status, see @ref OSReturnCodes @retval #OS_SUCCESS If any handle in the ReadSet or WriteSet is readable or writable, respectively @retval #OS_ERROR_TIMEOUT If no handles in the ReadSet or WriteSet became readable or writable within the timeout @retval #OS_ERR_OPERATION_NOT_SUPPORTED if a specified handle does not support select @retval #OS_ERR_INVALID_ID if no valid handles were contained in the ReadSet/WriteSet