leodos_libcfs::ffi

Function OS_SelectSingle

pub unsafe extern "C" fn OS_SelectSingle(
    objid: osal_id_t,
    StateFlags: *mut uint32,
    msecs: int32,
) -> int32

@brief Wait for events on a single file handle

Wait for a single OSAL filehandle to change state

This function can be used to wait for a single OSAL stream ID to become readable or writable. On entry, the “StateFlags” parameter should be set to the desired state (OS_STREAM_STATE_READABLE and/or OS_STREAM_STATE_WRITABLE) and upon return the flags will be set to the state actually detected.

As this operates on a single ID, the filehandle is protected during this call, such that another thread accessing the same handle will return an error. However, it is important to note that once the call returns then other threads may then also read/write and affect the state before the current thread can service it.

To mitigate this risk the application may prefer to use the OS_TimedRead/OS_TimedWrite calls.

The timeout is expressed in milliseconds, relative to the time that the API was invoked. Use OS_SelectSingleAbs() for higher timing precision.

@param[in] objid The handle ID to select on @param[in,out] StateFlags State flag(s) (readable or writable) @nonnull @param[in] msecs Indicates the timeout. Positive values will wait up to that many milliseconds. Zero will not wait (poll). Negative values will wait forever (pend)

@sa OS_SelectSingleAbs()

@return Execution status, see @ref OSReturnCodes @retval #OS_SUCCESS If the handle is readable and/or writable, as requested @retval #OS_ERROR_TIMEOUT If the handle did not become readable or writable within the timeout @retval #OS_INVALID_POINTER if argument is NULL @retval #OS_ERR_INVALID_ID if the objid is not a valid handle