leodos_libcfs::ffi

Function OS_SocketAcceptAbs

pub unsafe extern "C" fn OS_SocketAcceptAbs(
    sock_id: osal_id_t,
    connsock_id: *mut osal_id_t,
    Addr: *mut OS_SockAddr_t,
    abs_timeout: OS_time_t,
) -> int32

@brief Waits for and accept the next incoming connection on the given socket

This is used for sockets operating in a “server” role. The socket must be a stream type (connection-oriented) and previously bound to a local address using OS_SocketBind(). This will block the caller up to the given timeout or until an incoming connection request occurs, whichever happens first.

The new stream connection is then returned to the caller and the original server socket ID can be reused for the next connection.

This API is identical to OS_SocketAccept() except for the timeout parameter. In this call, timeout is expressed as an absolute value of the OS clock, in the same time domain as obtained via OS_GetLocalTime(). This allows for a more precise timeout than what is possible via the normal OS_SocketAccept().

@param[in] sock_id The server socket ID, previously bound using OS_SocketBind() @param[out] connsock_id The connection socket, a new ID that can be read/written @nonnull @param[in] Addr The remote address of the incoming connection @nonnull @param[in] abs_timeout The absolute time that the call may block until

@sa OS_SocketAccept()

@return Execution status, see @ref OSReturnCodes @retval #OS_SUCCESS @copybrief OS_SUCCESS @retval #OS_INVALID_POINTER if argument is NULL @retval #OS_ERR_INVALID_ID if the sock_id parameter is not valid @retval #OS_ERR_INCORRECT_OBJ_TYPE if the handle is not a socket @retval #OS_ERR_INCORRECT_OBJ_STATE if the socket is not bound or already connected