leodos_libcfs::ffi

Function OS_SocketConnect

pub unsafe extern "C" fn OS_SocketConnect(
    sock_id: osal_id_t,
    Addr: *const OS_SockAddr_t,
    timeout: int32,
) -> int32

@brief Connects a socket to a given remote address.

The socket will be connected to the remote address and port, if available. This only applies to stream-oriented sockets. Calling this on a datagram socket will return an error (these sockets should use SendTo/RecvFrom).

The timeout is expressed in milliseconds, relative to the time that the API was invoked. Use OS_SocketConnectAbs() for higher timing precision.

@param[in] sock_id The socket ID @param[in] Addr The remote address to connect to @nonnull @param[in] timeout The maximum amount of time to wait, or OS_PEND to wait forever

@sa OS_SocketConnectAbs()

@return Execution status, see @ref OSReturnCodes @retval #OS_SUCCESS @copybrief OS_SUCCESS @retval #OS_ERR_INCORRECT_OBJ_STATE if the socket is already connected @retval #OS_ERR_INVALID_ID if the sock_id parameter is not valid @retval #OS_ERR_INCORRECT_OBJ_TYPE if the handle is not a socket @retval #OS_INVALID_POINTER if Addr argument is NULL