leodos_libcfs::ffi

Function OS_SocketGetInfo

pub unsafe extern "C" fn OS_SocketGetInfo(
    sock_id: osal_id_t,
    sock_prop: *mut OS_socket_prop_t,
) -> int32

@brief Gets information about an OSAL Socket ID

OSAL Sockets use generated names according to the address and type. This allows applications to find the name of a given socket.

@param[in] sock_id The socket ID @param[out] sock_prop Buffer to hold socket information @nonnull

@return Execution status, see @ref OSReturnCodes @retval #OS_SUCCESS @copybrief OS_SUCCESS @retval #OS_ERR_INVALID_ID if the id passed in is not a valid semaphore @retval #OS_INVALID_POINTER if the count_prop pointer is null