leodos_libcfs::ffi

Function OS_SymbolLookup

pub unsafe extern "C" fn OS_SymbolLookup(
    symbol_address: *mut cpuaddr,
    symbol_name: *const c_char,
) -> int32

@brief Find the Address of a Symbol

This calls to the OS dynamic symbol lookup implementation, and/or checks a static symbol table for a matching symbol name.

The static table is intended to support embedded targets that do not have module loading capability or have it disabled.

@param[out] symbol_address Set to the address of the symbol @nonnull @param[in] symbol_name Name of the symbol to look up @nonnull

@return Execution status, see @ref OSReturnCodes @retval #OS_SUCCESS @copybrief OS_SUCCESS @retval #OS_ERROR if the symbol could not be found @retval #OS_INVALID_POINTER if one of the pointers passed in are NULL