leodos_libcfs::ffi

Function OS_SymbolTableDump

pub unsafe extern "C" fn OS_SymbolTableDump(
    filename: *const c_char,
    size_limit: usize,
) -> int32

@brief Dumps the system symbol table to a file

Dumps the system symbol table to the specified filename

@note Not all RTOS implementations support this API. If the underlying module subsystem does not provide a facility to iterate through the symbol table, then the #OS_ERR_NOT_IMPLEMENTED status code is returned.

@param[in] filename File to write to @nonnull @param[in] size_limit Maximum number of bytes to write

@return Execution status, see @ref OSReturnCodes @retval #OS_SUCCESS @copybrief OS_SUCCESS @retval #OS_ERR_NOT_IMPLEMENTED @copybrief OS_ERR_NOT_IMPLEMENTED @retval #OS_INVALID_POINTER if the filename argument is NULL @retval #OS_FS_ERR_PATH_INVALID if the filename argument is not valid @retval #OS_ERR_NAME_TOO_LONG if any of the symbol names are too long @covtest @retval #OS_ERR_OUTPUT_TOO_LARGE if the size_limit was reached before completing all symbols @covtest @retval #OS_ERROR if an other/unspecified error occurs @covtest