leodos_libcfs::ffi

Function OS_TaskCreate

pub unsafe extern "C" fn OS_TaskCreate(
    task_id: *mut osal_id_t,
    task_name: *const c_char,
    function_pointer: osal_task_entry,
    stack_pointer: osal_stackptr_t,
    stack_size: usize,
    priority: osal_priority_t,
    flags: uint32,
) -> int32

@brief Creates a task and starts running it.

Creates a task and passes back the id of the task created. Task names must be unique; if the name already exists this function fails. Names cannot be NULL.

Portable applications should always specify the actual stack size in the stack_size parameter, not 0. This size value is not enforced/checked by OSAL, but is simply passed through to the RTOS for stack creation. Some RTOS implementations may assume 0 means a default stack size while others may actually create a task with no stack.

Unlike stack_size, the stack_pointer is optional and can be specified as NULL. In that case, a stack of the requested size will be dynamically allocated from the system heap.

@param[out] task_id will be set to the non-zero ID of the newly-created resource @nonnull @param[in] task_name the name of the new resource to create @nonnull @param[in] function_pointer the entry point of the new task @nonnull @param[in] stack_pointer pointer to the stack for the task, or NULL to allocate a stack from the system memory heap @param[in] stack_size the size of the stack @nonzero @param[in] priority initial priority of the new task @param[in] flags initial options for the new task

@return Execution status, see @ref OSReturnCodes @retval #OS_SUCCESS @copybrief OS_SUCCESS @retval #OS_INVALID_POINTER if any of the necessary pointers are NULL @retval #OS_ERR_INVALID_SIZE if the stack_size argument is zero @retval #OS_ERR_NAME_TOO_LONG name length including null terminator greater than #OS_MAX_API_NAME @retval #OS_ERR_INVALID_PRIORITY if the priority is bad @covtest @retval #OS_ERR_NO_FREE_IDS if there can be no more tasks created @retval #OS_ERR_NAME_TAKEN if the name specified is already used by a task @retval #OS_ERROR if an unspecified/other error occurs @covtest