leodos_libcfs::ffi

Function OS_TaskFindIdBySystemData

pub unsafe extern "C" fn OS_TaskFindIdBySystemData(
    task_id: *mut osal_id_t,
    sysdata: *const c_void,
    sysdata_size: usize,
) -> int32

@brief Reverse-lookup the OSAL task ID from an operating system ID

This provides a method by which an external entity may find the OSAL task ID corresponding to a system-defined identifier (e.g. TASK_ID, pthread_t, rtems_id, etc).

Normally OSAL does not expose the underlying OS-specific values to the application, but in some circumstances, such as exception handling, the OS may provide this information directly to a BSP handler outside of the normal OSAL API.

@param[out] task_id The buffer where the task id output is stored @nonnull @param[in] sysdata Pointer to the system-provided identification data @param[in] sysdata_size Size of the system-provided identification data

@return Execution status, see @ref OSReturnCodes @retval #OS_SUCCESS @copybrief OS_SUCCESS @covtest @retval #OS_INVALID_POINTER if a pointer argument is NULL