leodos_libcfs::ffi

Function OS_TimeBaseCreate

pub unsafe extern "C" fn OS_TimeBaseCreate(
    timebase_id: *mut osal_id_t,
    timebase_name: *const c_char,
    external_sync: OS_TimerSync_t,
) -> int32

@brief Create an abstract Time Base resource

An OSAL time base is an abstraction of a “timer tick” that can, in turn, be used for measurement of elapsed time between events.

Time bases can be simulated by the operating system using the OS kernel-provided timing facilities, or based on a hardware timing source if provided by the BSP.

A time base object has a servicing task associated with it, that runs at elevated priority and will thereby interrupt user-level tasks when timing ticks occur.

If the external_sync function is passed as NULL, the operating system kernel timing resources will be utilized for a simulated timer tick.

If the external_sync function is not NULL, this should point to a BSP-provided function that will block the calling task until the next tick occurs. This can be used for synchronizing with hardware events.

@note When provisioning a tunable RTOS kernel, such as RTEMS, the kernel should be configured to support at least (OS_MAX_TASKS + OS_MAX_TIMEBASES) threads, to account for the helper threads associated with time base objects.

@note This configuration API must not be used from the context of a timer callback. Timers should only be configured from the context of normal OSAL tasks.

@param[out] timebase_id will be set to the non-zero ID of the newly-created resource @nonnull @param[in] timebase_name The name of the time base @nonnull @param[in] external_sync A synchronization function for BSP hardware-based timer ticks

@return Execution status, see @ref OSReturnCodes @retval #OS_SUCCESS @copybrief OS_SUCCESS @retval #OS_ERR_NAME_TAKEN if the name specified is already used @retval #OS_ERR_NO_FREE_IDS if there can be no more timebase resources created @retval #OS_ERR_INCORRECT_OBJ_STATE if called from timer/timebase context @retval #OS_ERR_NAME_TOO_LONG if the timebase_name is too long @retval #OS_INVALID_POINTER if a pointer argument is NULL