leodos_libcfs::ffi

Function OS_TimeToRelativeMilliseconds

pub unsafe extern "C" fn OS_TimeToRelativeMilliseconds(
    time: OS_time_t,
) -> int32

@brief Gets a relative time value from an absolute time

This function computes the number of milliseconds until the given absolute time value is reached in the system clock.

@note This is intended to ease transitioning from a relative timeout value to and absolute timeout value. The result can be passed to any function that accepts a relative timeout, to mimic the behavior of an absolute timeout.

The return value of this function is intended to be compatible with the relative timeout parameter of various OSAL APIs e.g. OS_TimedRead() / OS_TimedWrite()

@param[in] time An absolute time value

@returns Milliseconds until time value will be reached @retval OS_CHECK (0) if time is the current time or is in the past @retval OS_PEND (-1) if time is far in the future (not expressable as an int32)