leodos_libcfs::ffi

Function OS_TimedRead

pub unsafe extern "C" fn OS_TimedRead(
    filedes: osal_id_t,
    buffer: *mut c_void,
    nbytes: usize,
    timeout: int32,
) -> int32

@brief File/Stream input read with a timeout

This implements a time-limited read and is primarily intended for use with sockets but may also work with any other stream-like resource that the underlying OS supports, such as pipes or special devices.

If data is immediately available on the file/socket, this will return that data along with the actual number of bytes that were immediately available. It will not block.

If the file position is at the end of file or end of stream data (e.g. if the remote end has closed the connection), then this function will immediately return 0 without blocking for the timeout period.

If no data is immediately available, but the underlying resource/stream is still connected to a peer, this will wait up to the given timeout for additional data to appear. If no data appears within the timeout period, then this returns the #OS_ERROR_TIMEOUT status code. This allows the caller to differentiate an open (but idle) socket connection from a connection which has been closed by the remote peer.

In all cases this will return successfully as soon as at least 1 byte of actual data is available. It will not attempt to read the entire input buffer.

If an EOF condition occurs prior to timeout, this function returns zero.

@param[in] filedes The handle ID to operate on @param[out] buffer Storage location for file data @nonnull @param[in] nbytes Maximum number of bytes to read @nonzero @param[in] timeout Maximum time to wait, in milliseconds, relative to current time (OS_PEND = forever)

@returns Byte count on success or appropriate error code, see @ref OSReturnCodes @retval #OS_ERROR_TIMEOUT if no data became available during timeout period @retval #OS_ERR_INVALID_ID if the file descriptor passed in is invalid @retval #OS_ERR_INVALID_SIZE if the passed-in size is not valid @retval #OS_INVALID_POINTER if the passed-in buffer is not valid @retval 0 if at end of file/stream data