leodos_libcfs::ffi

Function OS_TimerAdd

pub unsafe extern "C" fn OS_TimerAdd(
    timer_id: *mut osal_id_t,
    timer_name: *const c_char,
    timebase_id: osal_id_t,
    callback_ptr: OS_ArgCallback_t,
    callback_arg: *mut c_void,
) -> int32

@brief Add a timer object based on an existing TimeBase resource

A timer object is a resource that invokes the specified application-provided function upon timer expiration. Timers may be one-shot or periodic in nature.

This function uses an existing time base object to service this timer, which must exist prior to adding the timer. The precision of the timer is the same as that of the underlying time base object. Multiple timer objects can be created referring to a single time base object.

This routine also uses a different callback function prototype from OS_TimerCreate(), allowing a single opaque argument to be passed to the callback routine. The OSAL implementation does not use this parameter, and may be set NULL.

The callback function for this method should be declared according to the OS_ArgCallback_t function pointer type. The timer_id is passed in to the function by the OSAL, and the arg parameter is passed through from the callback_arg argument on this call.

@note This configuration API must not be used from the context of a timer callback. Timers should only be configured from the context of normal OSAL tasks.

@sa OS_ArgCallback_t

@param[out] timer_id Will be set to the non-zero resource ID of the timer object @nonnull @param[in] timer_name Name of the timer object @nonnull @param[in] timebase_id The time base resource to use as a reference @param[in] callback_ptr Application-provided function to invoke @nonnull @param[in] callback_arg Opaque argument to pass to callback function, may be NULL

@return Execution status, see @ref OSReturnCodes @retval #OS_SUCCESS @copybrief OS_SUCCESS @retval #OS_INVALID_POINTER if any parameters are NULL @retval #OS_ERR_INVALID_ID if the timebase_id parameter is not valid @retval #OS_ERR_NAME_TOO_LONG name length including null terminator greater than #OS_MAX_API_NAME @retval #OS_ERR_NAME_TAKEN if the name is already in use by another timer. @retval #OS_ERR_NO_FREE_IDS if all of the timers are already allocated. @retval #OS_ERR_INCORRECT_OBJ_STATE if invoked from a timer context @retval #OS_TIMER_ERR_INTERNAL if there was an error programming the OS timer @covtest