leodos_libcfs::ffi

Type Alias CFE_ES_AppInfo_t

pub type CFE_ES_AppInfo_t = CFE_ES_AppInfo;

\brief Application Information

Structure that is used to provide information about an app. It is primarily used for the QueryOne and QueryAll Commands.

While this structure is primarily intended for Application info, it can also represent Library information where only a subset of the information applies.

Aliased Type

#[repr(C)]
pub struct CFE_ES_AppInfo_t {

    pub ResourceId: u32,
    pub Type: u32,
    pub Name: [i8; 20],
    pub EntryPoint: [i8; 20],
    pub FileName: [i8; 64],
    pub StackSize: u32,
    pub AddressesAreValid: u32,
    pub CodeAddress: u32,
    pub CodeSize: u32,
    pub DataAddress: u32,
    pub DataSize: u32,
    pub BSSAddress: u32,
    pub BSSSize: u32,
    pub StartAddress: u32,
    pub ExceptionAction: u8,
    pub Priority: u16,
    pub MainTaskId: u32,
    pub ExecutionCounter: u32,
    pub MainTaskName: [i8; 20],
    pub NumOfChildTasks: u32,

}

Fields
ResourceId: u32

< \cfetlmmnemonic \ES_APP_ID \brief Application or Library ID for this resource

Type: u32

< \cfetlmmnemonic \ES_APPTYPE \brief The type of App: CORE or EXTERNAL

Name: [i8; 20]

< \cfetlmmnemonic \ES_APPNAME \brief The Registered Name of the Application

EntryPoint: [i8; 20]

< \cfetlmmnemonic \ES_APPENTRYPT \brief The Entry Point label for the Application

FileName: [i8; 64]

< \cfetlmmnemonic \ES_APPFILENAME \brief The Filename of the file containing the Application

StackSize: u32

< \cfetlmmnemonic \ES_STACKSIZE \brief The Stack Size of the Application

AddressesAreValid: u32

< \cfetlmmnemonic \ES_ADDRVALID \brief Indicates that the Code, Data, and BSS addresses/sizes are valid

CodeAddress: u32

< \cfetlmmnemonic \ES_CODEADDR \brief The Address of the Application Code Segment

CodeSize: u32

< \cfetlmmnemonic \ES_CODESIZE \brief The Code Size of the Application

DataAddress: u32

< \cfetlmmnemonic \ES_DATAADDR \brief The Address of the Application Data Segment

DataSize: u32

< \cfetlmmnemonic \ES_DATASIZE \brief The Data Size of the Application

BSSAddress: u32

< \cfetlmmnemonic \ES_BSSADDR \brief The Address of the Application BSS Segment

BSSSize: u32

< \cfetlmmnemonic \ES_BSSSIZE \brief The BSS Size of the Application

StartAddress: u32

< \cfetlmmnemonic \ES_STARTADDR \brief The Start Address of the Application

ExceptionAction: u8

< \cfetlmmnemonic \ES_EXCEPTNACTN \brief What should occur if Application has an exception (Restart Application OR Restart Processor)

Priority: u16

< \cfetlmmnemonic \ES_PRIORITY \brief The Priority of the Application

MainTaskId: u32

< \cfetlmmnemonic \ES_MAINTASKID \brief The Application’s Main Task ID

ExecutionCounter: u32

< \cfetlmmnemonic \ES_MAINTASKEXECNT \brief The Application’s Main Task Execution Counter

MainTaskName: [i8; 20]

< \cfetlmmnemonic \ES_MAINTASKNAME \brief The Application’s Main Task ID

NumOfChildTasks: u32

< \cfetlmmnemonic \ES_CHILDTASKS \brief Number of Child tasks for an App