leodos_libcfs::ffi

Type Alias CFE_ES_BlockStats_t

pub type CFE_ES_BlockStats_t = CFE_ES_BlockStats;

\brief Block statistics

Sub-Structure that is used to provide information about a specific block size/bucket within a memory pool.

Aliased Type

#[repr(C)]
pub struct CFE_ES_BlockStats_t {
    pub BlockSize: u32,
    pub NumCreated: u32,
    pub NumFree: u32,
}

Fields
BlockSize: u32

< \brief Number of bytes in each of these blocks

NumCreated: u32

< \brief Number of Memory Blocks of this size created

NumFree: u32

< \brief Number of Memory Blocks of this size that are free