leodos_libcfs::ffi

Type Alias CFE_ES_MemPoolStats_t

pub type CFE_ES_MemPoolStats_t = CFE_ES_MemPoolStats;

\brief Memory Pool Statistics

Structure that is used to provide information about a memory pool. Used by the Memory Pool Stats telemetry message.

\sa #CFE_ES_SEND_MEM_POOL_STATS_CC

Aliased Type

#[repr(C)]
pub struct CFE_ES_MemPoolStats_t {
    pub PoolSize: u32,
    pub NumBlocksRequested: u32,
    pub CheckErrCtr: u32,
    pub NumFreeBytes: u32,
    pub BlockStats: [CFE_ES_BlockStats; 17],
}

Fields
PoolSize: u32

< \cfetlmmnemonic \ES_POOLSIZE \brief Size of Memory Pool (in bytes)

NumBlocksRequested: u32

< \cfetlmmnemonic \ES_BLKSREQ \brief Number of times a memory block has been allocated

CheckErrCtr: u32

< \cfetlmmnemonic \ES_BLKERRCTR \brief Number of errors detected when freeing a memory block

NumFreeBytes: u32

< \cfetlmmnemonic \ES_FREEBYTES \brief Number of bytes never allocated to a block

BlockStats: [CFE_ES_BlockStats; 17]

< \cfetlmmnemonic \ES_BLKSTATS \brief Contains stats on each block size