leodos_libcfs::ffi

Type Alias CFE_FS_FileWriteMetaData_t

pub type CFE_FS_FileWriteMetaData_t = CFE_FS_FileWriteMetaData;

\brief External Metadata/State object associated with background file writes

Applications intending to schedule background file write jobs should instantiate this object in static/global data memory. This keeps track of the state of the file write request(s).

Aliased Type

#[repr(C)]
pub struct CFE_FS_FileWriteMetaData_t {
    pub IsPending: bool,
    pub FileName: [i8; 64],
    pub FileSubType: u32,
    pub Description: [i8; 32],
    pub GetData: Option<unsafe extern "C" fn(*mut c_void, u32, *mut *mut c_void, *mut usize) -> bool>,
    pub OnEvent: Option<unsafe extern "C" fn(*mut c_void, u32, i32, u32, usize, usize)>,
}

Fields
IsPending: bool

< Whether request is pending (volatile as it may be checked outside lock)

FileName: [i8; 64]

< Name of file to write

FileSubType: u32

< Type of file to write (for FS header)

Description: [i8; 32]

< Description of file (for FS header)

GetData: Option<unsafe extern "C" fn(*mut c_void, u32, *mut *mut c_void, *mut usize) -> bool>

< Application callback to get a data record

OnEvent: Option<unsafe extern "C" fn(*mut c_void, u32, i32, u32, usize, usize)>

< Application callback for abstract event processing