leodos_libcfs::ffi

Type Alias OS_static_symbol_record_t

pub type OS_static_symbol_record_t = _bindgen_ty_3;

@brief Associates a single symbol name with a memory address.

If the OS_STATIC_SYMBOL_TABLE feature is enabled, then an array of these structures should be provided by the application. When the application needs to find a symbol address, the static table will be checked in addition to (or instead of) the OS/library-provided lookup function.

This static symbol allows systems that do not implement dynamic module loading to maintain the same semantics as dynamically loaded modules.

Aliased Type

#[repr(C)]
pub struct OS_static_symbol_record_t {
    pub Name: *const i8,
    pub Address: Option<unsafe extern "C" fn()>,
    pub Module: *const i8,
}

Fields
Name: *const i8Address: Option<unsafe extern "C" fn()>Module: *const i8