leodos_libcfs

Macro main

macro_rules! main {
    ($app_main_struct:ty) => { ... };
}

A macro to define the entry point for a cFS application.

This macro generates the required CFE_ES_Main C function, which serves as the official entry point for a cFE application. It links this entry point to the safe Rust application lifecycle managed by leodos-libcfs::cfe::es::app::start.

Example

use leodos_libcfs::cfe::es::app::{App, AppMain};
use leodos_libcfs::error::Result;
// Define the state for your application.
struct MyAppState { /* ... */ }

impl AppMain for MyAppState {
    fn init(app: &mut App) -> Result<Self> {
        app.register_for_events()?;
        // ...
        Ok(Self { /* ... */ })
    }

    fn run_cycle(&mut self, app: &App) -> Result<()> {
        // ...
        Ok(())
    }
}

leodos_libcfs::main!(MyAppState);