leodos_libcfs

Module psp

CFE Platform Support Package (PSP) interface.

The PSP provides the lowest-level abstraction layer, interacting directly with the hardware and board support package (BSP). These wrappers expose some of the more common and useful PSP functions to applications, but many are unsafe due to their low-level nature.

Modules

cds

CDS (Critical Data Store) interface.

eeprom

EEPROM (Electrically Erasable Programmable Read-Only Memory) interface.

exception

Safe wrappers for PSP exception log functions.

mem

Low-level PSP wrappers for memory information and access.

restart

PSP restart and reset functions.

time

Safe wrappers for PSP high-resolution time functions.

version

Wrappers for PSP version query functions.

watchdog

Wrappers for PSP watchdog timer functions.

Functions

flush_caches⚠

Flushes processor data or instruction caches for a given memory range.

get_processor_name

Returns the PSP-defined processor name.

status_to_string

Converts a PSP status code to its symbolic name.