leodos_libcfs::psp

Module mem

Low-level PSP wrappers for memory information and access.

Structs

MemRangeInfo

Information about a specific memory range defined in the PSP memory table.

Functions

get_cfe_text_segment_info

Returns the location and size of the CFE text segment.

get_kernel_text_segment_info

Returns the location and size of the kernel text segment.

get_mem_range

Retrieves one of the records from the PSP’s memory table.

get_mem_ranges

Returns the number of memory ranges defined in the PSP’s memory table.

get_reset_area

Returns the location and size of the ES Reset information area.

get_user_reserved_area

Returns the location and size of the user-reserved memory area.

get_volatile_disk_mem

Returns the location and size of the memory used for the cFE volatile disk.

mem_validate_range

Validates a memory range against the PSP’s memory map.